[IMP] composer: edit references with keyboard
when the user types in the composer and the focus is on a token that
could be a reference, allow the user to use the arrow keys and shift to
select a reference.

When editing the user has to first type a character. We do not implement
yet the F2 to "enter the edit range mode"

We have refactored the tests and the keyboard events between the grid
and the composer to deduplicate the logic of selecting the ranges as
well as to be able to quickly bind a key sequence to a function