[IMP] functions: Add env to the evaluation context
Evaluating a function might require a particular service (e.g. an
rpc service). Since services are usually accessible from the
environnement in an Owl application, we make it avalaible in
the evaluation context of any function evaluation

```js
functionRegistry.add("GET_REMOTE_DATA", {
 description: "Call a remote procedure to get data",
 compute: async function () {
   return this.env.rpc(...);
  },
  args: args``,
  returns: ["STRING"],
});

```

In particular, this will be usefull to evaluate the `PIVOT` function
in the Odoo integration.

Note: this is a breaking change since `getters` are no longer
accessible in the function evaluation with `this.getters` but
with `this.env.getters`.