[IMP] sidepanel: Toggle sidepanel
This commit introduces a new function in the spreasheet sub environnement
allowing to toggle a given sidepanel.

This will be usefull in Odoo where we introduce an icon in the topbar which
opens a sidepanel. Clicking a second time on the icon while the sidepanel is
open should close it.

Sidenote: we changed the parent component of all tests in `side_panel_tests.ts`
because it was more testing the helper test component `GridParent` than the
spreadsheet itself (the entire content of `openSidePanel` could be commented
without any test failing).

Task 2168366

Co-authored-by: fleodoo <fle@odoo.com>