[IMP] tokenizer: small optimizations
The tokenizer is a critical part of the updateCell process.  Every
single formula has to be tokenized, even if it shares the same structure
with another formula.

So, for sheets with a significant amount of formulas, improving the
tokenizer, even in small way, can have a non trivial effect.

This commit optimize some paths in the tokenizer.  Even though these
changes are small, it leads in a quick benchmark to an improvement from
10.5s to 9.8s in loading a sheet with 500k formulas.  Not really life
changing, but then, this is a cheap change.