[IMP] Range: create a range plugins
Managing and updating ranges is an issue currently in o-spreadsheet:
every plugin that uses references to cells in their data structure needs
to adapt those structures in a predictable way  when the shape of the
grid changes (by adding or removing colums and rows).

Duplicating this logic in every plugin is, well code duplication
Centralizing this logic, but applying it on the ranges of the different
data structure is not efficient because most of them store the references
in string (in the format Sheet!CxRx) and manipulating them is not efficient.

To solve all those issues, we have created the range plugin.
Its main goals are:
- managing the link between any reference to a cell or a group of cells
and every plugin
- adapt the ranges size when needed, efficiently
- restore the range in a Sheet!CxRx string format when needing to display it to a user
- notify the consumers of ranges (all the plugins and the formulas) when
a range changes so every plugin can take action and update itself if needed