[IMP] *: save formula normalized
In order to facilitate a certain number of processing, as well as to
speed up the startup of a larger sheet, we will now store the formula
with their structure independent from the references it uses:
a formula like `=sum(A1:A4, A5) + 3` will become
`{text: "=sum(|0|, |1|) + 3", dependencies: ["A1:A4","A5"] }`

This has the following benefits:
--------------------------------
- every formula that have the same text is compiled to the same byte code
- the text itself is independent of its dependencies, we can use
it as a lookup key without tokenizing and parsing it to know if there is
already a compiled version of this formula
- because the previous formula cache implementation already tried to
extract the dependencies, but at a much later stages, the compiler was
much more complex, and the execution context was too. This changed
allowed me to greatly simplify the compiler.

and the following drawbacks:
----------------------------
- for now, we are storing the formula in the normalized version, as
well as the original content of the cell --> it uses more disk space & memory
- I had to increase the CURRENT_VERSION of the data and make a migration

closes #596