[REF] *: Split core and UI getters
This commit splits the getters into two parts: CoreGetters and Getters.
CoreGetters are all the getters defined on core plugins.
The Getters are all these getters, and all the getters defined on UI
plugins.

With that, core plugins no longer have access to the UI getters
(activeSheet, ...).
This prevents core plugins to depends on UI getters.
The same security should probably be done for commands.

Co-authored-by: pro-odoo <pro@odoo.com>