[REF] functions: remove some optional flags
Regarding function definitions:
Repeated arguments and default arguments are always set with the
'optional' flag.

This commit removes the 'optional' flag in cases where the 'repeating'
or 'default' flags already exist.

As a result, this commit changes the way the 'repeating' and 'default'
flags should be understood when evaluating arguments.