[REF] formatting: move styles and format to cell
Styles and formats are currently imported and exported by the cell plugin.
Also they are stored directly on cells...in the cell plugin.
Hence, it makes sense to move the full responsability of styles and formats to the
cell plugins.

This commit moves style and format management from the formatting to
the cell plugin. The formatting plugin now only handles borders and will
be renamed in a following commit.

Co-authored-by: LucasLefevre <lul@odoo.com>