[IMP] evaluation : formulas can return a format
This commit adds the possibility to define formats on formula functions.
So when rendering cells the format of the final formula is returned.

This is made possible because this commit allows when compiling a
formula to define a priority list of formats according to the functions
present in the formula.

The priority list is evaluated when creating a cell after compiling.

Note the case where the formula format depends on reference: when we
change the format of the reference, the format of the formula isn't
automatically updated.