[FIX] translation: correctly translate dynamic content
Using backtick to translate will not work with the translation system.
```js
const x = 2;
_lt(`Hello ${x}`);
```
`Hello ${x}` is evaluated *before* the lookup in the table of
translation, so we are looking for "Hello 2" in the table, which is
not correct.

This commit introduces the way to translate such strings in the following
way:
```js
const x = 2;
_lt("Hello %s", x.toString());
```