[IMP] plugins: Simplify command result
The current `CommandResult` is verbose. It does not help reading and writing
`allowDispatch` methods.

With this commit, the result is no longer a `status` "CANCELLED" or "SUCCESS"
with possibly a `CancelledReason`. Both status are unified under one single enum.
`allowDispatch` now returns an enum item. Which can be `CommandResult.Success`
or any other item (previoulsy the `CancelledReason`) which are considered as
failures.