[FIX] sheet: check the activeSheet is still valid in finalize
Messages that comes from outside (collaborative) are not dispatched in
the UI plugins (only `finalize` is called).

So, before this commit, loading a sheet with a command which delete the
activeSheet did not correctly update the activeSheet as it did not go
through `beforeHandle` of `SelectionPlugin`.

This commit moves the behavior of changing the activeSheet in the
`finalize` of `SelectionPlugin`.
It also ensure that `finalize` is called after the initial messages.

Co-authored-by: LucasLefevre <lul@odoo.com>