[FIX] gridcomposer : resize height in some cases
This commit solves display problems in the gridcomposer when the cell
height is bigger than the standard cell height.

1st problem: writing a formula that is too long results in a line break
with too large line spacing.

2nd problem: the formula assistant is displayed below the cell, whereas
it must be displayed below the text in the composer.

Now that dimensions are managed with CSS, tests verifying dimensions by
checking specific functions call have been removed.