[IMP] helpers: Remove optional toZone parameter
The `toZone` helper currently has an optional parameter `keepBoundaries`.
It makes it impractical to use with `map` because the map callback
receives the index as second argument [1].

e.g. `xcs.map(toZone)` is not correct and yet it's probably the first idea which
comes to mind.
The first xc (index=0) is converted with `keepBoundaries=0` which is falsy while
all others with stricly positve integers which are truthy.
The correct way would be `xcs.map((xc) => toZone(xc))`.

If the developer doesn't perfectly know the `map` API or he/she is sleepy (which
I would be on a late winter friday afternoon), it may lead to unexpected
behaviors and bugs which would not be straight forward to find.

Better to be safe and remove the optional parameter.
A new helper is introduced to cover the `keepBoundaries=true` use-case.

[1] https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Array/map#parameters