[IMP] tests: Snapshot canvas screenshot
Writing renderer tests is cumbersome. It relies on a mock canvas and tests
must check what properties/methods the renderer calls. It is hard to write
and read.
As a result, the renderer is poorly tested.

Moreover, this testing strategy is fragile since it relies on the renderer
implementation.

This commit addresses both issues at once.
We can now "screenshot" the canvas and save it as a jest snapshot.
Like regular jest snapshot, tests will fail if the screenshot is not
identical to the previous one.

Rendering tests are now easy to write :)
There no more excuse not to write them!

Small notes:

- Rendering use Roboto font or aria if it is not found on the system. We
don't want those tests failing based on installed fonts. Therefore we must
inclued our own font files.

- PNG images are committed in the repo. It shouldn't be much of a problem
since they would usually be small. But makes sure to only screenshot what
you need ;)