[FIX] selection: clip all sheet selections with undo and redo
FORWARD PORT:
When forward porting this, please see the cleaner implementation and tests in #834

Steps to reproduce

On sheet1, add a column and then select the last column.
Change sheet (should already exist)
Undo
Go back to sheet1 => traceback
After an undo or redo operation, the current selection may not be valid
anymore. It is currently correctly udpated for the active sheet.
However, the saved selections of other sheets are not updated.

fixes #763