[IMP] clipboard: do not always update formulas on CUT/PASTE
During a cut-paste operation, some formulas do not need to be updated,
depending on the target of the ranges.

A range contained in the source zone should be updated, while a range
which is outside the source should be updated.

Let's take the following example:
A1: 2
B1: =A1

1) Select B1, CUT & PASTE to C1, the formula should not be updated.
2) With the initial situation, Select A1:B1, CUT & PASTE, the formula
should be updated.

Closes #856