[IMP] spreadsheet: allow running spreadsheet in readonly mode
This commit introduces a new parameter to instantiate a spreadsheet
component: isReadonly.
This flag is used to indicate if the spreadsheet is running in readonly
mode (i.e. the user cannot update the document).

To ensure that the user cannot alter the document, we need to refuse
the commands he run, but not all. We need to keep some of them, like
selection, evaluation, sheet activation, ...
The way to define a such command is explained in the documentation.