[REF] model: check if core command is allowed from UI plugins
Currently, when a UI plugin dispatches a core command, it is unconditionnaly
handled by core plugins (no `allowDispatch`). This means that it's the
responsability of the UI plugin to dispatch correct/sensible core commands
that wouldn't lead to crashes or a corrupted state.

This sometimes required to implement some checks before dispatching. Those
checks are often similar to what's implemented in the core `allowDispatch`
methods.

With this commit, `allowDisptach` is executed before core plugins handle
commands dispatched by a UI plugin.