[IMP] plugins: return multiple cancell reasons from allowDispatch
This is a prelimiary work for task 2342152

In some cases we want to know all reasons for which a command is refused by
`allowDispatch` because we want a visual feedback in the UI for all of them.
e.g. if multiple values are missing or invalid in the command, we want the
corresponding inputs to provides hints of the error.

Task 2342152