[FIX] evaluation: evaluate empty styled cells as empty
Currently, if A1 only has a style but no content, it is
not correctly recognized as an empty cell. Referencing
it (i.e. with the formula `=A1`) should give 0 but it
gives an empty string.

Since the introduction of cell types in 67ca3a7, an empty cell
can no longer be recognized by an empty `content`. The key "content"
is not even a valid key in the `Cell` type.

Co-authored-by: rrahir <rar@odoo.com>