[FIX] Charts: fix Pie charts display
Currently, Pie charts support several datasets but the colors are not
share between them of a same label.
Furthermore, The tooltip on the graph doesn't allow to distinguish the
different datasets.

This commits addresses both issues by creating a fixed color set for
each dataset and by adding a custom toolip callback to Chart.js options.