[IMP] conditional_formatting: do not reset conditional formatting when the type is changed
Before this commit, the current edition was lost when switching the type
of CF.

This commit also clean and simplify the side panel of CF and the editors
of the rule. The main side panel is responsible of the CF without the rule,
which is the responsibility of the editors.

Odoo-task-id 2532335