[FIX] sheet,evaluation: Support references to sheet names containing a space
The main issue was actually fixed by 0b7eaad
This commit adds some tests and some cleaning.

With this commit, quotes are also displayed around cross-sheet references in
the SelectionInput for sheet names that need it. Note that it is only for
consistency and "cosmetic". It correctly works without quotes (probably by
chance).

Co-authored-by: Lucas Lefèvre <lul@odoo.com>

Task 2556824