[FIX] selection: handle selections inside hidden headers
Currently, once a user-selection ends up (partially) hidden due to some
sheet manipulation, any attempt to change this selection with keyboard
inputs - which is synonym of dispatching `ALTER_SELECTION` *with a delta* or
`MOVE_POSITION` - will end up in some of these situations:
- move the position of  a hidden selection in the direction of the hidden header will
fall in an infite loop,
- alter_selection doesn't take hidden headers into account -> it will
not "jump" a hidden header while extending the selection
- some moves (e.g. select A#, hide col A , trying to move position to
the left) will just leave the selection unchanged, while leaving it
nearly invisible to the user's view.

This commit aims to prevent any infinite loops and introduces some
fallback behaviour so the user can regain some control of his selection.
mainly:

- A selection is frozen if both its columns and rows are hidden,
- We can now dispatch a move_position as long as the destination cell is
not hidden
- We can alter_selection as long as the current selection containe one
visible header in the alter direction (I.e. a visible column for a
vertical move and a visible row for a horizontal move)
- For both commands, trying to move in a direction will select the first
visible in the given direction. If it can't find any, it will start
looking in the opposite direction. Since we prevent a sheet to have all
its columns/rows hidden, we are guaranteed to find a candidate, given
that we are not in the situation of the first point.

Furthermore, in case a selection could end up hidden, the user can only
rely on the autofill bottom square to identity his position on the grid.
When the sheet is longer than the viewport and that the selection is a
complete hidden column, the autofill suare is just not visible. The user
has no clue of where his selection is but keyboard inputs might still
alter it, which can cause a lot of confusion to the user.

To improve the user experience, this commit allows hidden elements to
be drawn in the canvas. They will appear as lines (compressed
rectangles) but will sill provide a "pointer" on the user's location.