Alternative implementation
This commit tries to take advantage of the scrollbar properties to
define the actual maximum offset of the viewport. Instead of a
theoretical limit that is basically the maximum offset we can readh to
have at least one col and 1 row visible in the viewport, we use the
maximum scroll value of the scrollbar, which is clearly the maximum
offset that the client should be able to reach.