[FIX] viewport,grid: prevent scrolling loop
The new viewport implementation introduced a loop when drawing the grid.
It could trigger a dispatch to update the viewport offset, which in turn
would trigger a render, thus a `drawGrid`, etc...

The issue is the following:
- We scroll which a mousewheel, edge-scroll, ...;
- this will dispatch `SET_VIEWPORT_OFFSET` - Since we have a snapping
viewport, this operation might alter the offset we actually store inside
the plugin;
- the dispatch ends up with a rerendering;
- Once rendered, we redraw the grid;
- When drawing, since the viewport might have altered the offset, **we
set the scrollbars offsets with the valule stored in the plugin**.

If the values differ, then we redispatch `SET_VIEWPORT_OFFSET`.

The values will only differ on specific navigators (i.e. Chromium-core
navigators) for specific zomm level and [devicePixelRatio](https://developer.mozilla.org/en-US/docs/Web/API/Window/devicePixelRatio)
because in those navigators, the scrollbars will only accept values
(scrollTop and scrollLeft) that are multiples of the inverse of the pixel
ratio*. This means that the scrollbar might only accept non-integer
values while the viewport plugin might have stored an integer value
while adjusting the offset to snap the viewport. Combine this situation
with the fact that trying to set floating values is prompt to rouding
errors, we might restart the flow several times. to the point where the
last scrolling set will differ enough from the original scroll that the
viewport zone is altered.

In that situation, when edgescrolling, which will basically try to override
the scrolling offset over and over again, we are stuck in a loop where
the viewport will jump back-and-forth.

A part from that zooming/devicePixelRatio issue, itlooks like we were
dispatching `SET_VIEWPORT_OFFSET` twice quite often, rendering the grid
twice for no valid reason.`

This commit prevents the possibility of dispatching `SET_VIEWPORT_OFFSET`
when drawing the grid.

\* E.g. if the devicePixelRatio (dpr) is 1.1, the chrome will accept
steps of 1/1.1 ~= 0.909091
Trying to set a given sive will lead to the following results:
input 0,5 -> result 0
input 1   -> result ~0.909091
input 2   -> result ~1.818182

With rounding issues, we might see

input ~1.818182 => result ~0.909091 becausewith the rounding, the value
provided is sllightly below  2 * 1/1.1