[ADD] data_source
DataSource is an abstract class that contains the logic of fetching and
maintaining access to data that have to be loaded from outside.

A class which extends this class have to implement two different methods:
* `_fetchMetadata`: This method should fetch the metadata, i.e. data that
should be fetch only once.
* `_fetch`: This method should fetch the data from the server.

To get the data from this class, there is three options:
* `get`: async function that will returns the data when it's loaded
* `getSync`: get the data that are currently loaded, undefined if no data
are loaded
* specific method: Subclass can implement concrete method to have access to a
particular data.

Each DataSource is managed into the DataSourcesRegistry. Its role is to
ensure that an evaluation is triggered when a data source is ready and
to provide a way to ask all the data source to be loaded.
This registry is available in the config object, given to all plugins.

Odoo-task-id 2607786