[FIX] core: Reference update on header deletion
Currently, when deleting a column (resp. row), references are updated
based on the assumption that they span over several columns (resp.
rows).
When the reference spans over a single column (resp. row) (e.g. A1:A2),
and that that specific header is being deleted, the "right" part of the
reference is shifted to the left of the "left" part of the reference,
making it invalid. When this happens on column A (resp row 1), the code
crashes because we try to process an invalid zone (where right= -1).

e.g.

in C1,
input "=sum(A1:A2)"
delete column A
=> a traceback occurs

Task 2619342