[IMP] edition: avoid changing all highlights colors when adding ref.
In edit mode when we add or remove the reference of a cell (or range)
in the composer before another reference, this modifies the colors of
all the following references.

This problem also occurs when we move or resize a highlight on a
highlight with the same reference, all the following highlights
change colors.

This commit aims to avoid that.