[FIX] evaluation: correctly handle cells in error state
Before this commit, the cells in error were not reset before a new
evaluation. So, if a cell positioned before another cell in error mode, and
the given cell depends on the cell in error, the evaluation detected the
dependency as an error and thus mark the given cell as error.

Task-id 2634778