[FIX] position hook: update .o-spreadsheet position on mounted
In Odoo, the menu on a chart figure is wrongly positioned.
It is shifted down be the height of the nav bar and the control panel.

The reason is because the `useAbsolutePosition` hook gets the reference
position of `.o-spreadsheet` before the spreadsheet is mounted
(it falls back to {0, 0}) and it's never updated once the spreadsheet is
mounted.