[REF] tests: remove helper component GridParent
The component `GridParent` used in tests has grown to an
half-duplicated implementation of the `Spreadsheet` component.

Small comment about adapted tests:
The `GridParent` component set the viewport size to the size
of its `el`, which contains the top bar. However, the real
implementation set the viewport size to the grid size.

closes #808