[IMP] selection: drag and drop selection on the grid
Add a new component named "selection" allowing to cut/paste a
selection by grabbing and moving it.

Note that a new selection event is added to make the difference
between:

1 - a new zone selected
2 - a zone moved

In the first case, selecting a new zone automatically update
the viewport to make the anchor zone still visible.
In the second case, the viewport is not automatically updated.

Making this difference allows to update the viewport according
to the location of the cursor and not the location of the zone.

task-id: 2560650