[FIX] composer: Ensure legitimate rendering of topbar composer content
Since 59c6de8a1fc8507e8ac15352d653af4e987f0abe, we rely on a
`isKeyStillDown`to avoid untimely re-rendering inside the composer input
box.
This boolean is only resetted when a `keyup`event is captured but
unfortunately, such event does not occur when we process the Enter key
or when we lose focus while pressing a key (pressing a key then click
anywhere on the grid). The topbar composer was then stuck in a state
where we wouldn't update the content anymore unless we'd refocus it and
trigger a `keyUp`event on it.

Part of task 2647122