[REF] types: automatically type getters
This commit improves the type safety of getters. It also
reduces the required boilerplate when adding a new getters.

Prior to this commit, you could add in the `getters` list:
- the name of a private method, it would work
- the name of a method which does not exist. It would crash at runtime, not
  at compile time.
- a method name, but reference another method when declaring it in the Getters
  type. e.g. `aMethod: CellPlugin["anotherMethod"];`

With this commit, it's no longer necessary to manually add new getter methods
to the `Getters` type. Adding the method name to the `getters` static attribute
is enough.

Task 2676799