[FIX] edition/selection_input: component loss of focus was not reflected on the plugins
When switching focus between the composer and a selection_input component,

Form a functional pov, the focus on the composer and a selectionInput
component are mutually exclusive (it's also the case between 2 seletioninput
components).

This was not supported at the plugin level and we could end up with the
focus on a selectionInput while the edition mode was active. This led to
faulty behaviour where the Highlight (resize and move) component was set
for a seletionInput while it only support the composer highlights right now.

Furthermore, the highlights of both the composer and the selection input
were displayed at the same time.

Task 2677071

Co-authored-by: Lucas Lefèvre <lul@odoo.com>