[FIX] selection_input: fix single range inputs
When selecting the ranges of a single range input (namely, the label
range of a chart), the extension of a selection was not properly
supported and we could still add several ranges[1]. A part of the code
was assuming that only one range would exist which would lead to crashes
[2].

Right now, we only use two kinds of inputs, the ones with a single
range and those with unlimited number of ranges. This commit adapts the
code to that specific situation and drops the possibility of having
limited amounts of ranges (different than 1).

[1] - focus your input
    - key Ctrl pushed (= extending the selection)
    - Add several cells at once
   => you now have multiple ranges in your labelRange input.

[2] - Once [1] is done and you have several ranges in labelRange, switch
to the 'Design' tab, then back ton 'Configuration'.
    - Only one range is visible
    - Repeat [1] but by adding several zones of multiple cells
   => there is a crash at the rendering level.

Task 2677909

Co-authored-by: Lucas Lefèvre <lul@odoo.com>