[FIX] Composer: ensure correct rendering of the composer
This commit addresses two issues:

1. the state of the function descriptor was reset with
a static object whenever calling the method `processTokenAtCursor`which
was breaking the `useState` hook. Because of this, the function
descriptor was not rerendered when adding a character that was modifying
the edition mode to `waitingForRangeSelection`. The isseu dissapeared
when a rerendering occured.

e.g. Type '=sum(1',  then type ',' the function descriptor should be
highlighting the second argument.

2. When processing the keyUp event in the composer, we did not
necessarily dispatch a command or trigger a rerendering, even if the
content of the input was altered. This means that we didn't update the
content style, i.e. not add the indicator for range selection. This
commit forces a refresh of the content.

e.g. Type '=', then delete the character and retype it: the caret
indicator is not present.

Task 2698146

Co-authored-by: Lucas Lefèvre <lul@odoo.com>