[REF] *: remove ui actions in plugins
This commit prepares the remove of t-ref on components (2703524).
The main idea is to initialize the model outside of Spreadsheet. It will
feel more natural and we could avoid accessing the model from the parent.

To do that, the first step is to simplify the config of the model.
One thing that is not clean is to call ui actions (edit text, ask
confirmation, ...) directly from the model.
Instead, we move this behavior directly in the components, which can
react however they want based on the result of allowDispatch.
The ui actions have been moved from the model to the env, which is
available on all the components.

It allows to remove the interactive mode of a command, which leads to
strange behavior: bypass allowDispatch, ...

Now, a plugin has only one way to contact the UI: `notifyUI`.

Part of task-id 2703524