[REF] cf: editors are now basic templates
This task is a step to make spreadsheet owl2 compatible.

Before this commit, the three editors (cell is rule, color scale and icon set)
were components.
This design came with some issues:
* To be able to react to input changes, the CF should be observed by the
  editor => we made something like this.state = this.props !
* To be able to save a CF, the main component should ask it child its
  current state => bad communication way

Now, the editors are now basic templates with only one component which
is responsible of the state of the CF.

Task-id 2713118