[FIX] clipboard: correctly selected pasted zone outside sheet
Copy a zone and past it at the very bottom of the sheet.
It adds the missing rows. But the selection is now wrong.

Here is what happens:
- it select the (soon to be) pasted zone
- it adds missing rows (which shift the selection)
- it paste values

Adding rows adapts the current selection (because of collaborative),
but the selection was already set to the pasted zone as if the rows
already existed. To prevent such mistakes in the future, it's now
impossible to set the selection outside the sheet: the selection
is clipped automatically.