[FIX] normalize: Normalize number and string arguments in functions
While the normalization of ranges is a drastic improvement of
performance, as the strings and numbers are not taken in account, we
still keep compiling a lot of formulas which call the very same function
but don't have the exact same arguments.

e.g.  `=PIVOT("1", "group_field", "5")` and `=PIVOT("1", "group_field", "6")`
are compiled 2 times. Actually, any combination of arguments in a pivot
function will be recompiled.

This commit introduces the normalization of strings and numbers
arguments. Booleans had to be left aside at this time but since they
only have two values, they are less likely to make the number of
compiled functions explode.

Note: Inside a sheet containing 732 000 cells with pivot functions,
we went from 732000 compiled formulas to 16.

Task 2720073

Co-authored-by: Pierre Rousseau <pro@odoo.com>