[FIX] evaluation:function error for range outside of sheet
SUM(AA1) (with AA1 outside of sheet) had an error at the evaluation.
The expected behaviour is to have =0.

This happened because the _range function of evaluation.ts threw an error
when receiving a range outside of the sheet.

Now the _range function returns an array with undefined instead of throwing
an error.

Odoo task 2722593