[FIX] viewport,grid: prevent scrolling loop
The new viewport implementation introduced a loop when drawing the grid.
It could trigger a dispatch to update the viewport offset, which in turn
would trigger a render, thus a `drawGrid`, etc...

The issue is the following:
- We scroll which a mousewheel, edge-scroll, ...;
- this will dispatch `SET_VIEWPORT_OFFSET` - Since we have a snapping
viewport, this operation might alter the offset we actually store inside
the plugin;
- the dispatch ends up with a rerendering;
- Once rendered, we redraw the grid;
- When drawing, since the viewport might have altered the offset, **we
set the scrollbars offsets with the valule stored in the plugin**.

If the values differ, then we redispatch `SET_VIEWPORT_OFFSET`.

This commit removes the loop by avoiding to assign the scroll value at
render. To do so, we take advantage of the `eventBus` of `Model` to set
the scrolling value of the scrollbars which will then dispatch a
`SEVT_VIEWPORT_OFFSET` that will ultimately trigger a render.

Task 2620080

closes odoo/o-spreadsheet#1108

Signed-off-by: Pierre Rousseau (pro) <pro@odoo.com>
Co-authored-by: Lucas Lefèvre <lul@odoo.com>