[REF] spreadsheet: move clipboard handling to grid
This commit is a step to the owl 2 compatibility. A t-ref is still there
in spreadsheet to reference grid. A use case for this was to check if
the active element is grid in order to handle copy/cut/paste.

In order to avoid this, we can move the handling of the clipboard to
grid directly.

Part of task id 2733428