[FIX] collaborative: drop local revisions after empty commands.
This issue is a bit tricky. Let's begin with the use case:
1) A command is created by Alice
2) Concurrently, Alice undo her command, and Bob do a command that is
valid with the command of Alice, but not anymore without. (Ex: insert
a cell in a sheet that was created by Alice. If the command of Alice
is removed (aka undo-ed), the sheet does not exist anymore).
3) Alice redo her command.

At this point, the command of Bob would be valid as the sheet is here.
But, when Bob try to send his command, the command is empty (have been
transformed with the inverse of CREATE_SHEET). So, if he sends his
command, Alice will not be able to insert his command in all of her
branches, because obviously it's not possible to retrieve the
SET_CELL_CONTENT from an empty command.

To solve this, the behavior is to drop a command (and all the following)
is the command is *local* and is empty.

Co-authored-by: LucasLefevre <lul@odoo.com>