[FIX] collaborative: compute inverse transformation with last received commands
The inverse transformation is not computed with newly received commands
taken into account.

The following bit explains the test added in this commit.
Hopefully it will help the reader to understand.

Here is Charlie's situation before receiving Bob's command.
His own local command is D. And the transformed version D(4) is transformed
with the inverse of B (noted B⁻¹) in the upper branch.

  STAR(0)   A(1)   B(2)   C(3)   D(4)
                   >      C(5)   D(6)

  D(4) = D(6) t.w. B(2)⁻¹  where t.w. stands for "transformed with"

Now Charlie receives command X which must be inserted before D. The transformed version D(5)
is not correctly computed. The upper branch should always reflect the lower branch without
the undone command. Which means all commands in the lower branch transformed with B⁻¹.
The order of transformations matters!

  STAR(0)   A(1)   B(2)   C(3)   X(4)   D(5)
                   >      C(6)   X(7)   D(8)

Currently, D(5) is computed as:
  D(5) = D(4)             t.w. X(4)
       = D(6) t.w. B(2)⁻¹ t.w. X(4)

The inverse transformation is still performed on D(6) which does not
reflect the current lower branch (with Bob's command taken into account)

It should be:
  D(5) = D(8)           t.w. B(2)⁻¹
       = D(6) t.w. X(7) t.w. B(2)⁻¹