[FIX] demo: fix function `makeLargeDataset`
When introducing commit ea578bc08bf1f7e2a675380559a9edb5dbc89499,
we changed the format of imported data by denormalizing it. This change
was not reflected in the function `makeLargeDataset`.

The function was also improved to allow one to generate sheets
containing string and number values to test performances in wider cases.

task 2749784