[FIX] collaborative: export cell styles in the same order
When exporting cells, the iteration order depends on key insertion order.
In a collaborative context, it may happen that several users locally inserted
some cell in a different order, resulting in a different iteration order.
Since style ids are auto incremented ids, their value could be different.

It doesn't make any difference in practice, but to keep things consistent and
allow to compare exported data in tests this commit ensures all users export
cells and styles in the same order.