[FIX] evaluation: limit re-evaluations triggered by datasources
Currently, datasources are able to notify that they have updated their
data and each of these notification will trigger a full evaluation of
the current sheet.

Unfortunately, we encounter situations where the number of notifications
explodes.
E.g. in [1], each call to this function will trigger an evaluation. Said
evaluation could handle a `PIVOT.HEADER` formula, which will in turn
trigger an evaluation if the cache is not updated yet, etc ...

Concretely, if the triggers occur in batch, only the last one is relevant.
This commit therefore adds a debounce to the event 'data-loaded' triggered
by the dataSource registry, which will  in turn drastically reduce the number
of evaluations.

In case of [1], the number of evaluations is now equal to the number
of groupBys in the pivot.

[1]: https://github.com/odoo/enterprise/blob/15.0/documents_spreadsheet/static/src/js/o_spreadsheet/plugins/ui/pivot_structure_plugin.js#L84-L97

task 2752589