[IMP] edition: display current number with percent format
Previous commit improved the way you can edit percentages for new content.
Setting a new number in a percent formatted cell displays the number with
a %

However, when editing the *current* content the number is still displayed
in decimal: 3% is displayed as 0.03
This is inconsistent: new content is displayed in percent but current in
decimal.

With this commit, the composer content of a percent formatted cell is a
percent representation.

Task 2766311