[FIX] components: Adjust css rules
The css rules of certains components were not precise enough and only
worked because all css rules were declared in a given order. They lacked
a correct specificity.
i.e. the rules of component `Spreadsheet` were declared before the rules
of its children, thus in case of conflicting properties (here
`box-sizing`) the rules of the children prevailed.

Since we moved to owl2 and the STYLE static property was removed from
the component, we had to come up with a new way to declare the css rules
and the order of declaration was altered. The style òf `Spreadsheet`is
now declared after the style of its children components and the rule of
`Spreadsheet` prevails.

This commit changes the selectors to be more precise whenever the
properties were conflicting to make sure we apply the property of the
child.

See https://developer.mozilla.org/en-US/docs/Web/CSS/Specificity#:~:text=When%20multiple%20declarations%20have%20equal%20specificity%2C%20the%20last%20declaration%20found%20in%20the%20CSS%20is%20applied%20to%20the%20element

Task 2767810