[FIX] viewport: Allow valid non-integer offsets
The commit 2c25af1b05fecd66aced6b15069dd1679d587a4f was not properly
adapted when forwardported in `saas-15.1`.

For starters, see  the original issue (here)[https://www.odoo.com/web#id=2620080&action=333&model=project.task]

The issue is that the client can sometimes dispatch a non integer offset
value, which starts to be an issue at the maximum scroll value.
There is a discrepency between the maximum value that can be set on a
scrollbar (potentially non-integer) and the one we get by computing
`getMaxViewportSize() - viewportWidth/height`. The former can be bigger
than the latter and that is why we have to consider a (+1) in the
`allowDispatch`.

This flow is not testable because inside jest, we cannot alter the
`deviceDisplayRatio` and scrollbars will only take integers offset
values.

note: Technically, we had to stop clipping the offset in
`setViewportOffset` because it prevented us from scrolling completely.

Task 2768982