[REF] evaluation: 60% faster evaluation
With this commit, evaluation is up to 3 times faster.
On the "large formula dataset"
with 200k cells, the evaluation time goes from ~450ms to ~150ms.
with 500k cells, from 1200ms to 500ms
with 50k cells, from 120ms to 55ms

I'm not sure why though.
A performance analysis with Chrome dev tools doesn't seem to show that
`getCellPosition` and `toXC` are particularly slow

closes odoo/o-spreadsheet#1222

Signed-off-by: Pierre Rousseau (pro) <pro@odoo.com>