[IMP] charts: add scorecard and gauge charts
Add scorecard chart
-------------------

Add "Scorecard" chart type. Create new getters in chart.ts and
evaluation_chart.ts specific for this type of chart.

The chart itself is not a ChartJS chart, but is a standard Component
with hand-written logic for scaling the font sizes.

Some work should be done at a latter date in the chart plugins to better
handle all types of chart and avoid the repeated if/else.

Add gauge chart
---------------

Gauge chart based on "chartjs-gauge" library

Revamp chart management
-----------------------

Since the addition of scorecard and gauge chart, it becomes a mess to
manage different types of charts. There is a duplication of many concepts,
like getters (getBasicChartDefinition, getGaugeChartDefinition, ...),
local state (charts, scorecard, gauge), ...
In addition, it was not possible to add a new type of chart from outside
o-spreadsheet (for example add Odoo charts).

This commit introduce something similar to cellFactory => a chart Factory.
Based on the type of the chart in the definition of a chart, an instance of
a Chart is created and stored in the local state of chart core plugin.
A Chart is an instance of a class (which inherits from AbstractChart),
which implements some functions that will be called from the plugin.

For each Chart class, a Chart Runtime class should be created to
compute the runtime data needed by the components.
A Side panel should also be created (a config panel and a design panel).

Tasks 2796686, 2796685 and 2868625

closes odoo/o-spreadsheet#1393

Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>
Co-authored-by: Pierre Rousseau <pro@odoo.com>
Co-authored-by: Adrien Minne <adrm@odoo.com>
Co-authored-by: Alexis Lacroix <laa@odoo.com>