[IMP] link: open link popover when hovering cell
Instead of opening the o-link-tool popup when the cell is selected:
- opening it when the cell is hovered for more than Xms
- close it when the cursor stay stationary for more than Xms outside of the o-link-tool popup

With this commit, the popovers link/error are no longer managed
individually by the grid, but are managed by a plugin `CellPopoverPlugin`.

To create popover, you can either :
- use the `OPEN_CELL_POPOVER` command
- add a matcher in the `cellPopoverRegistry`. The matcher takes as argument
a position of the grid, and return a popover component if a popover should
be opened when hovering this position.

Task 2637513

closes odoo/o-spreadsheet#1187

Signed-off-by: Pierre Rousseau (pro) <pro@odoo.com>
Co-authored-by: Adrien Minne <adrm@odoo.com>