[REV] sort: move back sort command from core to UI
This reverts commit f6ea663

The sort command handling depends on the evaluated cell values.
The values are not guaranteed to be the same for every users because
they depends on: asynchronous rpc calls, user accesss rights.

Handling the same sort command can therefore lead to different results for
different users which would be de-synchronized.

Task 2801574

Part-of: odoo/o-spreadsheet#1272