[REF] sheet: simplify sheets getters
Before this commit, we had two misnaming getters:
* getSheets         Return the Sheet instances of the ordered sheets
* getVisibleSheets  Return ths Sheet ids of the ordered sheets

It's necessary to clarify the concept of "visible sheet" before
introducing the feature of hide/show sheets.

So, visibleSheets has been renamed by orderedSheets.
While renaming it, it appears that getSheets was not really necessary and
all the use cases can be covered with getVisibleSheets.

So, instead of the two getters, there is only one remaining:
`getOrderedSheetIds` which returns the ids of the ordered sheets.

Part-of: odoo/o-spreadsheet#1302