[FIX] highlights: Handle merges correctly
Steps to reproduce:
- type =SUM(A1:A10), look at the highlight
- merge A2:B2
- click on the SUM, the highlight is wrong, it covers A2:B10 instead of
  A1:A10.

After investigation on other tools (GSheet & Excel) the rules seems to
be the following:
When the range to highlight covers a single cell and that it is part of
a range, then we highlight the whole merge. The rest of the time, we
only highlight the cells covered by the range.

E.g.

Consider a merge in A1:C3,

A1     -> highlight A1:C3
C3     -> highlight A1:C3
A1:A10 -> highlight A1:A10
A1:A2  -> highlight A1:A2

task 2833434

closes odoo/o-spreadsheet#1310

Signed-off-by: Pierre Rousseau (pro) <pro@odoo.com>