[IMP] config: purge dist files when switching branches
We can sometimes forget to purge our dist folder when switching branches
which can be deceiving when running human tests (i.e. testing a demo
sheet).
This can sometimes lead to awkward bugs or behaviour that we
believe to detect which are just residuals of another branch/build.

Just as we have to reset our npm packages it seems fair to do the same
thing with the dist folder.

Task 2837807

closes odoo/o-spreadsheet#1328

Signed-off-by: Pierre Rousseau (pro) <pro@odoo.com>