[IMP] spreadsheet: increase cell font resize its row
When the user increase the font size of a cell, the row size should be updated,
except if the row had previously been resized by the user.

Don't export sizes or rows that were automatically computed, but re-compute them at import.

This commit also fix AUTORESIZE_ROW to align its behaviour with GSheet. Now
the minimum size possible for a row auto-resized is the default cell height.

Odoo task 2714374

closes odoo/o-spreadsheet#1359

Signed-off-by: Pierre Rousseau (pro) <pro@odoo.com>