[FIX] drag&drop: Use the correct delay to edgeScroll
in the `dragAndDropBeyondTheViewport` helper, we would use the time delay
computed for a horizontal edgescroll when scrolling vertically. This
meant that when scrolling vertically, we would call the
`onMouseMove`callback with a timeout delay of 0ms, thus scrolling way
too fast.

Writing a test for this particular case highlighted the complexity of
the necessary setup to test such a feature as well as the fact that we
need to test it on EVERY component that can edgeScroll, and the list of
seuch components can only grow over time.

A work targetting the master branch will try to address this issue by
testing the actual helper in a more generic context, which should in
turn hopefully allow to make simpler tests upon the components themselves.

task 2850618

closes odoo/o-spreadsheet#1362

X-original-commit: a83dc7b7cebb66b5343f5d79ae3c22c42520917f
Signed-off-by: Pierre Rousseau (pro) <pro@odoo.com>
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>