[IMP] errors_handling: add a log level to display errors
TASK DESCRIPTION

The aim of this task is to hide the ERROR symbole (red triangle in
the upper right corner of the cell) for some kind of error.
This is done by adding an ErrorLevel to the error classes so that
the errors having 'silent' as level are not displayed.

Currently, only the NotAvailable error is silent, as this error
is defined by the user.

RELATED TASK

task-2868276

closes odoo/o-spreadsheet#1392

Signed-off-by: Pierre Rousseau (pro) <pro@odoo.com>