[FIX] config: remove `.gitattributes` file
Continuation of 658553f4. The added `.gitattributes` file was a terrible
idea and should have been thoroughly tested accross stable branches
priorly. The declaration was such that all files (including binaries)
were affected by the change. This caused the very few images we have to
be corrupted when we checkout them.

Since this config is not essential, it is better to remove it
altogether.

closes odoo/o-spreadsheet#1412

X-original-commit: ba941bcfd0a65d5bb3b2e637089eb01d49a99c43
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>