[IMP] renderer: Hide gridlines over color formatted cells in dashboard
Currently, we draw gridlines  over cells that have a colored background.
This should only happen under 2 conditions - the grid lines are visibles
(per user choice) and we are not in dashboard mode.

From a technical POV, the rest of the time, we draw a border in the same
color as the cell background.

Task 2883210

closes odoo/o-spreadsheet#1441

Signed-off-by: Pierre Rousseau (pro) <pro@odoo.com>