[FIX] spreadsheet.ts: fixed show formula bug
Earlier there was a bug with 'search and replace' side panel, that
upon clicking show formula button and closing the side panel directly
would not change the state of formula visibility, resulting with
formula being displayed on cells instead of value.

To fix this issue, an event is dispatched while closing the side panel
which will set the formula visibility as per settings done via view menu.

Task-2882964

closes odoo/o-spreadsheet#1446

Signed-off-by: Pierre Rousseau (pro) <pro@odoo.com>