[FIX] scorecard: chart background above figure borders
During resizing of a scorecard, the background color was displayed above
the figure borders.

This commit change the props of the charts from `figureId` to `figure`.
This means that we have access to the current figure size during the
resizing of a chart instead of the saved figure size.
We can thus update the size of the scorecard dynamically, and avoid its
background to be above the figure borders.

Odoo task 2906393

Part-of: odoo/o-spreadsheet#1491