[IMP] notification: Add new notification type
Recent tasks have shown the need for a new kind of notification to the
user.

The current type (notifyUser) has been developped/used for the sole
purpose of warning the user when a given action could not be performed.
It implies that this notification should stop the user in his work, to
make sure he acknowledges the issue he faced. In htis sense, it is not a
real notification.
Such 'notification' is ATM only triggered by the user's own action, as it
would be troublesome to be blocked by the actions of some other user.

However, we could (and will, in the near future) need less intrusive
notification; so that the user could be warned without being blocked.
Such notification would have to be non-blocking; simply informative;
 an actual notification.

This commit adds a new function raiseError to address the current use
of notification (taht is warn the user upon a forbbiden action) and
alters the purpose of notifyUser to be non-invasive.

Part of task 2696839

Part-of: odoo/o-spreadsheet#1457
Co-authored-by: Anthony hendrickx <anhe@odoo.com>