[FIX] evaluation: re-evaluate errors in other sheets
Cell in errors in other sheets are not re-evaluated.
The error is rethrown right away when they are referenced.
This is not correct since formulas result may have changed
because of an asynchronous computation.

The issue is similar to the issue fixed in 35b67d3, but
with other sheets.

Part-of: odoo/o-spreadsheet#1509