[FIX] evaluation: re-evaluate async formulas when changing sheets
Let's say a cell was previously evaluated to a given value,
then the user switch to another sheet. If the formula result changes
in the meantime (because of an async computations), the formula is not
re-evaluated (because it's not the active sheet).
When going back to the sheet, the user will still see the previous value.

closes odoo/o-spreadsheet#1514

X-original-commit: a64059db949825b831a4865f63833d2d7446e4d3
Signed-off-by: Pierre Rousseau (pro) <pro@odoo.com>
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>