[FIX] overlay: persisting handle after header resizing
When resizing an header, an handle appears and the cursor change to a
double arrow. There was a problem that after a double click (which
cause an auto-resize), the handle persisted in the middle of nowhere.

Now update the state of the handle after a double click even in the
overlay.

Odoo task 2928598

closes odoo/o-spreadsheet#1538

Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>