[FIX] evaluation: formula cell evaluated.value as null
Before this, it was possible that the evaluated.value of a formula cell
was null. This happened if the formula was a simple reference to an
empty cell. The evaluated value was thus null, but the formatted value
was "0". This lead to a discrepancy between what the user see (a "0"),
and the internal value of null.

This was problematic for example when evaluating CFs, the user see a
cell "0", but internally it's null thus not matching a CF "equal 0".

This commit make it so formula cells cannot have an empty evaluated value.
If the formula return "null", it's converted to 0.

Odoo task 2867122

closes odoo/o-spreadsheet#1578

Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>