[FIX] figures: positions are not wysiwyg
We had a problem where the figures position was (0;0), but the figures
were clearly not displayed in (0;0).

This happened because of a CSS `left:3px;top:3px;` in the figures. This
was added because of the bad structure of the figure wrapper HTML :
the wrapper had `overflow:hidden` to avoid having the figure overlapping
with the headers. At the same time, the anchors to move the figure were
naturally overflowing beyond the figure border, so the solution to display
them was to have an offset of 3px to always be able to display them
without overflow.

The fixed structure of the container is (Bottom-Up):

- o-figure: this div contains the actual chart. It has a fixed size
that don't depend on the scroll offset, and its overflow will be cut by
its parent. The borders need to go on this div to be cut with the chart.

- o-figure-wrapper: wraps the figure to avoid overflows. Has the same
size as its parent.

- o-figure-container: root element that contains the wrapper and the
anchors. We resize and shift it to only display the visible part of the
figure (with regard to the scroll offset).  We need an additional offset
to move it when the figure is selected since the border of o-figure
changes size. The div allows overflows to display the anchors correctly.

Odoo task 2543173

closes odoo/o-spreadsheet#1579

Signed-off-by: Pierre Rousseau (pro) <pro@odoo.com>