[FIX] test: mock grid size for link display
Test for the link display didn't mock the grid size.
When the code was trying to resize the viewport, it didn't properly described
reallity because it was dispatching an invalid command with a negative size
(rejected by an allowDispatch validation).
The size was: grid.clientWidth - SCROLLBAR_WIDTH which results in a negative
number if the clientWidth isn't mocked and always is 0

Without properly mocking the grid size, the next commit would break some tests
because the command is no longer invalid since we don't remove the scrollbar and
the viewport size would be set to 0

Part-of: odoo/o-spreadsheet#1630