[IMP] evaluation: add error "unknown function"
This commit add a new error type UnknownFunctionError at evaluation to
replace the generic bad expression error.

This error is handled at parsing, when the parser encounter an unexpected
SYMBOL token, it will throw a `UnknownFunctionError` instead of a generic
error if the symbol is a possible function.

A SYMOBL is a possible function if it match a regex for function names,
and the next token is an `OPEN_PAR`.

Odoo task ID : 2824111

closes odoo/o-spreadsheet#1617

Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>