[REF] components: use dedicated component for dashboard mode
Currently, the dashboard mode is rendered with the regular Spreadsheet and
Grid components, with many branching conditions for dashboard specificities
(if (this.env.isDashboard()) { ... }).
It's a beautiful mess.
This commit introduces a dedicated Dashboard component. Now the Grid component
is independent from dashboard mode.
Many common things needed to be factorized to avoid duplicated code.

Going further in the future: we still use Spreadsheet component which has many
useless code for dashboard mode. We could stop using Spreadsheet and directly
use Dashboard by adding the necessary props and env related things.
But it's less an issue because it has way less branching condition in
Spreadsheet (only one)

Implemention comment about useGridDrawing:
Drawing the grid must be the last thing done. Especially it must be done after
the sheet view has been resized according to the screen. One might think
resizing the view (by a dispatched command) would re-render the grid anyway
but it's not true for the first render! The event handler that rerender the
spreadsheet at each model update is only attached when Spreadsheet component
is mounted (onMounted). But it's the very last to be mounted since it's the
app root component.
The solution chosen here is to have onMounted hook in the parent of
GridOverlay (which resizes the grid) to make sure it's executed after resizing.
It's factorized in useGridDrawing to avoid duplicated code.

X-original-commit: 9dc58781fdea760899fb9831462f9232ceada66c
Part-of: odoo/o-spreadsheet#1686
Co-authored-by: Pierre Rousseau <pro@odoo.com>