[FIX] interactive merge: remove useless user confirmation
If the user try to create a merge that both overwrite a value and is inside
a data filters, the user will be prompted to confirm overwriting values,
then the merge will fail because it's created inside a data filter. But
no error message will show up.

Fix this by displaying the MergeInFilter error before prompting the user.
It's useless to ask confirmation from the user for something that won't work.

Odoo task 2977521

closes odoo/o-spreadsheet#1816

X-original-commit: a17522c8a3164229bc7cd473276b3060344b4135
Signed-off-by: Pierre Rousseau (pro) <pro@odoo.com>