[FIX] topbar: dropdown closing when clicking inside
There was a bug where clicking inside a color picker/border dropdown
closed the dropdown even when clicking between the clickable elements.
Clicking inside the dropdown shouldn't close it.

The dropdown was actually closed by 2 separate mechanisms:

- the dropdown was a child of the tool button that opens/closes the
dropdown. Thus clicking on the dropdown called the click event of the
tool and closed the dropdown.
- the topBar uses a externalListener on clicks, and this closed the
dropdown too.

Fixed the issue by :

- making the dropdown content a sibling of the button that opens it rather
than a child. This means it won't call the button onClick, and don't inherit
css, like cursor style, from it.
- stopping the propagation of click events in the dropdown. The onClick
of the items in the dropdown now have the responsibility to close it.
Odoo task ID : 2937081

closes odoo/o-spreadsheet#1564

Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>