[FIX] composer: display raw content of NumberCell
Before this, the composer displayed the value of a number cell with the
default format.

This was a problem when typing a number with a lot of decimal in the grid.
The last few digits were formatted away to keep the number short, which
is nice to have in the grid, but since it also happened in the composer
the last few digits were totally lost to the user.

Added a test for this problem.

Odoo task 2977643

Part-of: odoo/o-spreadsheet#1733