[IMP] cell: Remove useless uuid generation for performance
The cells ids are generated using random UUIDS but those are only
'runtime', they are not stored in the exported data and not used in any
command whatsoever. Generating the uuids is a bit costy and in this
specific case, can be replaced with an incremental id generator.

Benchmark:
----------

Copy/pasting 26k cells:

With *uuid*: 4.95 sec

With *incremental id*: 2.23 sec

Note that this is not an issue at start up since `uuidv4()` is shortcut
to an incremental id generator at that moment.

All-in-all,  the uuid generator only makes sense for persistent ids.

Task 2987517

closes odoo/o-spreadsheet#1738

Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>