[FIX] data filters: traceback at sheet duplication
There was a traceback at the duplication of a sheet containing data
filters, but only if the sheet was imported with the data filters, not
if they were freshly created.

This happened because the state of `filter_evaluation` wasn't correctly
initialized at import, and there was a `?.` missing.

Odoo task 3034403

closes odoo/o-spreadsheet#1743

X-original-commit: c256ceb83247a52af11a5ce94e85d241f7f3ae8b
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>