[FIX] selection: update default selection anchor on dispatched commands
How to reproduce:
-----------------

From a spreadsheet with 2 sheets:
- reload for a fresh start
- on Sheet1, select cell B4
- delete sheet1
- (Now by dfault on sheet2, your selection is A1)
- hit `ArrowRight`
=> the selection moves to C4 instead of B1

After a command that would alter the spreadsheet and impact the
selection, the default selection anchor was not properly updated in the
stream processor, which would lead to faulty behaviours as the one
described above. Took the opportunity to purge a useless function.

part of task 3032860

closes odoo/o-spreadsheet#1740

closes odoo/o-spreadsheet#1770

Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>