[FIX] selection: Altering selection should scroll the viewport
Commit b437f446b introduced an error. while fixing the viewport
scrolling while in autofill (it shouldn't happen) it also prevent an
change of selection (not an extension) to properly scroll the viewport.

The problem ultimately lies in the fact that the autofill was using the
selectionProcessor to override its selection instead of extending it.

In order to fix this, it was necessary to alter the API of the
selection processor to allow more granularity in the steps to extend a
zone selection: i.e. instead of a step of 1 or jump to the end of a zone,
we need to support all POSITIVE values of steps.

Task 3050101

closes odoo/o-spreadsheet#1823

X-original-commit: 765150d4d25157d2f7afb9885a6a93213eaec664
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>