[FIX] figure: prevent duplication of figure id
Currently, one can duplicate a figure id which is a  two-fold problematic.

1. Anyone could overwrite an existing figure with another if they were
to redispatch the same sheetId/id  combination to create a figure.
Thankfully, it's not the case ATM

2. The problem already exists when duplicating charts. In ChartPlugin, the
mapping only relies on the figureId and not on the sheetId. Unfortunately,
we are using an incremental id used on sheet duplication to ensure the
synchronization in multiuser (e6c1c369). This id is reset when we
reload a new model and we can create situations where we overwrite a
chart entry when duplicating a sheet.

A more in-depth rework of the Figure/chart mapping is required in master

Task 2999075

X-original-commit: a66f517387de60bc8aa40b000da608bcd401d6d6
Part-of: odoo/o-spreadsheet#1801