[REF] models: separate ui plugins in three concepts
Before this commit, plugins were split in two: core and ui, based on
whether a state should be shared in collaborative.

With this commit, we split the ui plugins in three categories:
- Plugins which have a state, but which should not be shared in
  collaborative (edition, selection) => stateful_ui_plugins
- Plugins which have a derived state from core data (evaluation) =>
  core_views_plugins
- Plugins which handle a specific feature, without handling any core
  commands (autofill, sort) => feature_plugins

For now, the only difference is that a command replayed will be dispatched
to core view plugins (typically to ensure the evaluation is correct even
after a command replay).

But in the future, we should give the history to stateful ui plugins, in
order to save the selection or the active sheet when a core command is
dispatched.

Part of task-id 2813749

closes odoo/o-spreadsheet#1817

Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>