[REF] selection: getPosition return main position
Basically every call to `getPosition` is followed
by a call to `getMainCellPosition` (or should be).
It is easy to forget and cumbersome

The getter `getPosition` now returns the "main position"
i.e. the top-left coordinate of the merge if the current
selection anchor is within a merge.

Task 3076136

Part-of: odoo/o-spreadsheet#1819