[IMP] model: config to provide external dependencies
This commit introduce a proper and unified way to provide external
dependencies to plugins and functions.

It's currently possible for functions[1] but not for plugins.

Now you can provide external dependencies in the model config by
adding a `external` key which can hold any arbitrary value.

[1] The current way to add external dependencies to functions is to
add a custom evaluation context in the model context (`evalContext`).

This was hijacked in odoo to access external dependencies in plugins
even if they are unrelated to evaluation.

```ts
class CustomPlugin extends UIPlugin {

    constructor(config) {
        this.env = config.evalContext.env;
    }
}
```
now becomes
```ts
class CustomPlugin extends UIPlugin {

    constructor(config) {
        this.env = config.external.env;
    }
}
```

Part-of: odoo/o-spreadsheet#1831