[IMP] Error: Lookup module's functions  now return NA Errors
Task Description

When using a function of the LOOKUP module, whe should get a N/A error
when the value we are looking for doesn't exist in the specified range.
To be able to get error that make sense anyway (as the NA error is silent
nowadays), we then added the possibility to put a custom error message when
throwing a NA error. If a custom message is added, the error will have the
classical error level, while it will have the silent error level in the
other case.

Related Taks ID
task-3067622

closes odoo/o-spreadsheet#1844

Signed-off-by: laa-odoo <laa@odoo.com>