[IMP] renderer: Long centered contents overflowing and overlapping on neighoring cells
Overflowing text-centered cells would sometimes not be cropped properly
when their content would overlap on a neighbor* cell.
The problem was that the condition to apply the clipping was not
considering the position at which the text was centered but only the
content width compared to the size of the space the said content could
expend over. If the text was overflowing on the neighbor* right
cell but we could extend sufficiently to the left, we wouldn't
consider the need to clip, hence the overlapping of text on the right.

* here a neighbor is not the direct neighbor, this specific case was
already handled. THe isue would occur when writing in C1 and having the
text overlap over E1 for example.

task 3088107

closes odoo/o-spreadsheet#1847

Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>