[REF] config: move data external dependencies to config
Currently, some external dependencies are injected in three
different ways to the spreadsheet application

Direct user interactions: in the sub-env
file store: `<Spreadsheet/>` props
others: config.external

The goal of this commit is to reduce the number of different ways to
inject those dependencies.

With this commit, all "data related" dependencies are injected
in the `config`.
Namely: `fileStore` and `loadCurrencies`

`config.external` is kept for custom things that o-spreadsheet doesn't
care about (additional plugins, functions, etc.)

Task 3093071

Part-of: odoo/o-spreadsheet#1863