[IMP] components: Remove sidePanelIsOpen props from components
This props was forwarded all the way down to the figures so they could
register themsleves as the content of the side panel. However, this was
not the right way to handle it for several reasons:

1. This feature only made sense to change the side panel content when a
user would alter selecting figures. The implementation here would
display the figure side panel even though another type of side panel was
already open,e.g. Conditional formats or Search and Replace.

2. The chart side panel already handles this itself since it updates
itself when the user selects a new figure.

The props is still passed down to the `Grid` component as it is required
to define the grid size which also has an impact on the plugins. It will
however disappear once we reorganize the global css of the components
( see https://github.com/odoo/o-spreadsheet/pull/1769)

task 3087981

closes odoo/o-spreadsheet#1866

Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>