[REF] edition: force consistency in the plugin state
The plugin internal state is globally inconsistent. Some values are
properly resetted when start/stopping the edition, some are not.
Furthermore, functions are designed such that we need to call fucntions
in a very specific order in order not to crash [1].

This commit forces a stronger typing in the plugin.

[1] https://github.com/odoo/o-spreadsheet/pull/1739#discussion_r1041233112