[IMP] selection_input: consistent range colors
Current behaviour of range colors upon being focused in selection input
is it shows different colors every time the sidepanel is opened while
range is being focused.

To bring a consistency on these colors, simply an index parameter is
passed in colors array in order to remove the global variable and the
helper method that is maintained for color index.

Task-2877044

closes odoo/o-spreadsheet#1890

Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>