[FIX] Composer: Persistent composition when starting the edition
When inputing a character in the grid, we capture the character and use it
to start the edition, which opens a new composer and focus it. This
change of focus closes the Input Method Editor if present. This
typically affect people from Asia writing in their native language
(chinese, japanese, etc)

This revision addresses the issue by replacing the hidden input in the
grid with the `GridComposer`. The later will therefore always be in the
DOM, capture the characters with the IME and then change its position
based on the edition mode.

closes odoo/o-spreadsheet#3619

closes odoo/o-spreadsheet#3628

Task: 3685891
X-original-commit: acc8569
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>