[FIX] charts: render only if needed
Create a chart with a lot of data (e.g. 20k cells)
=> scrolling doesn't feel smooth when the chart is visible.

Rendering a chart JS is a costly operation.
Before this commit, it was re-rendered on every owl rendering,
even if the data did not changed.

with 20k cells:
before: 200-300ms
after: not perceivable

The cost is still perceivable when the chart enters the viewport.

closes odoo/o-spreadsheet#1925

Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>