[FIX] sheetView: grid composer at wrong pos when cell not in viewport
Grid composer renders at 0,0 location on sheet when cell is not
in viewport, main reason seems to be ActiveViewport returns 0,0 when
cell is out of viewport.

To fix this issue, upon START_EDITION refreshViewport will be triggered
which will auto scroll the sheet to cell until it's in viewport.

Task-3081142

closes odoo/o-spreadsheet#1973

X-original-commit: 32c036b92d9580912ac274529fd45f6389ecd2cb
Signed-off-by: Pierre Rousseau (pro) <pro@odoo.com>