[FIX] grid_overlay: properly detect grid overlay resize
Before be10c889, the opening/closing of the side panel would update the
props of `GridOverlay`, which in turn would trigger the useEffect that
updated the plugin sheetview size. Unfortunately, the aforementioned
commit removed that props, which means `GridOverlay` was no longer
patched.

This commit proposes to rather rely on the web API `ResizeObserver`
whose role is to report the change of size of any observed `Element`.
This will act as an external listener, hence we don't need for
`GridOverlay` to be rendered in order to acknowledge its change of size
in the browser.

Task 3125767

closes odoo/o-spreadsheet#1977

X-original-commit: 0fb68dec310091a8988346cf9b2cf0f23cd0b5c0
Signed-off-by: Pierre Rousseau (pro) <pro@odoo.com>