[IMP] autofill: support edge scrolling
This commit support the edge scrolling in autofill.
The basic idea is to reuse `dragAndDropBeyondTheViewport` function.
Besides it, we also need to cancel the check of whether the autofill
is active (which is based on the visibility of the autofill selection)
to make the tooltip and handler still available during edge scrolling.

task 2870875

closes odoo/o-spreadsheet#1961

Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>
Co-authored-by: rrahir