[FIX] data filters: don't copy filter values
If a user duplicates a sheet and another user concurrently creates
a data filter in that sheet, it crashes if the sheet duplication is
first.

The filter evaluation thinks the filter was copied on the new sheet but
it actually didn't because the data filter didn't exist when the sheet
was duplicated.

It happens because remote revisions are dispatched to UI plugins *after*
everything is dispatched to core and the data filter actually exists
in the first sheet.

Task 3131564

X-original-commit: 93a28d03917a605006b23f4418e78aded16dd3e9
Part-of: odoo/o-spreadsheet#2015