[FIX] header size: compute the size before duplicating the sheet
When duplicating a sheet A, the header sizes are copied as well.
However, they should be copy in the current state of sheet A.

Since the values are computed lazily, we need to make sure they are
computed.
In particular, before this commit, if the original sheet was deleted,
the lazy computation would crash because the lazy closure was still
referencing sheet A.

Task 3131564

X-original-commit: 3777c192db98499eccce4033f548608ba29b2cae
Part-of: odoo/o-spreadsheet#2015