[FIX] topbar: properly toggle menus
The 'toggling' of the menus was not really doing what it said because it
would always end up setting up new menu items and would never close the
menus. The functional behaviour was still good though be only because
the menu component has an external listener that will make it close when
we click outside the menu.

This revision introduces a real toggling which will spare some useless
instanciations of `Menu` and renderings.

Part-of: odoo/o-spreadsheet#1985