[FIX] scorecard: rendering issue while resizing
Since 5514296, the rendering of the scorecard was bugging a bit during
resizing, the title was blinking in and out.

This was because the size of the scorecard was updated during the
`useEffect()` which changed the state of the chart, thus we had to wait
for the next render to have a correct value.

In this commit we removed the borders in the `o-figure` element. This
means that we can use the size of the figure props in scorecards.

The figure borders are now set in the `o-figure-border` overlay above
the figure. This mean that 1px in the chart is now covered by the border,
which shouldn't be a big problem.

Odoo task 3130899

closes odoo/o-spreadsheet#2074

X-original-commit: 0bd1eeafb50dc7742e6b62e40fa620f33bdfd603
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>
Signed-off-by: Minne Adrien (adrm) <adrm@odoo.com>