[FIX] popover: preserve scroll in popover
Currently, each new rendering will reset the popover style which will
also reset its scrollTop(or Left) property. This is an issue that can
already be noticed when opening a menu too big for the screen. when the
said menu will be rerendered (for instance when opening one of the
submenus), the scroll is resetted and the mouse is now targetting
another submenu, making the first one unreachable.

Task 3146014

closes odoo/o-spreadsheet#1985

Signed-off-by: Pierre Rousseau (pro) <pro@odoo.com>