[FIX] components: Fix static colors for spreadsheet
Some components had part of their style relying on bootstrap defaults.
Those can be altered by changing css variables, to create a dark mode
for example.

The issue is that our custom style cannot be alteredin the same way
because we rely on JS variables to generate static css whereas we should
rely on scss variables to use in js. Since the current implementation
does not support scss variables, we force the light mode for the time
being.

Task 3143440

closes odoo/o-spreadsheet#2032

X-original-commit: cd1f96ae72f090e55b146d696792baf7884d6565
Signed-off-by: Pierre Rousseau (pro) <pro@odoo.com>
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>