[FIX] Renderer: Speed up rendering
The rendering process has been slower since the introduction of the
frozen panes (c0c2044). One of the main reason being that we were
computing the boxes sizes and positions in the past but now have to do
it for multiple panes at a time. This is done by calling the getter
`getVisibleRect` for each box zone. However, we were also calling
`getSheetViewVisiblecols/rows` to get the viewport  boundaries for each
cell, it would take between 25 and 39% of the call to `drawGrid`.

This revision ensures that we only fetch the viewport boundaries once
when computing the boxes.

closes odoo/o-spreadsheet#2029

Signed-off-by: Pierre Rousseau (pro) <pro@odoo.com>