[REF] tests: introduce click helper
Before this revision, we were using different ways to click on elements
(dispatching an event by hand, using our helper triggerMouseEvent).
When the event was dispatched by hand, we were not always using the
bubbles option, which could lead to some issues.

This revision introduces a new helper click that allow us to easily
click on an element. It uses the MouseEvent constructor and correctly
set the bubbles option to true.

Task-id 3165489

closes odoo/o-spreadsheet#2028

Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>