[REF] tests: Use minimal setup for composer child components
the tests of `TextValueProvider` and `FunctionDescriptionProvider` AKA
autocomplete_dropdown and formula_assistant; were instantiating a full
spreadsheet for their tests even though the only needed to be starting
from a composer.

This revision extracts the composerParent from the composer tests in
order to reuse it for the composer children components.

closes odoo/o-spreadsheet#1998

Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>