[IMP] evaluation: introduce evaluation by round
With the incoming spreading matrix formulas, we may have formulas spreading a matrix
able to depend on a matrix spread from another formula.

As the current evaluation process was not enough to handle cross-spread dependencies, we
needed to introduce a new evaluation scheme. We decided to make evaluation per round, meaning
that every updated cell during the evaluation process will trigger the re-evaluation of all
its references in the next evaluation round. We then go from one round to another as long as
we still have cell to recompute.

This new scheme has been needed to be able to handle these cross-dependencies without knowing
the size of the future spread before the evaluation, which is something that won't be available
without evaluating the formula.

Unfortunately, as all values should be computed inside its round, we had to remove the lazy evaluation
process.

Co-authored-by: Anthony Hendrickx (anhe) <anhe@odoo.com>