[REF] figure: unify chart and image menus
Before this commit, the chart and image menus were handled separately in
`figure_chart.ts` and `figure_image.ts`. This led to a lot of duplicated code
since they both do 80% the same thing.

Now the figure menu is handled in `figure.ts`, and a `menuRegistryBuilder`
was added to `figureRegistry`.

Odoo task 3199007

closes odoo/o-spreadsheet#2091

Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>