[IMP] history: add the possibility to repeat commands at redo
Now if we do a redo and the redo stack is empty, we will try the redo
the last played command if possible. The list of repeatable commands
is defined in the repeat_commands_registry.ts file.

Repeated commands are transformed to match the current selection/sheet.
To be able to repeat local commands, the "root" command of a revision
was added to the history.

Locals commands of stateful UI plugins cannot be simply transformed and
repeated, because the state of The plugin may have changed. There is 2
strategies to solve this :

1. If possible, transform core sub-commands instead (eg. STOP_EDITION)
2. If not possible, we have to store the last state of the plugin, and
dispatch a new commands REPEAT_* (eg. REPEAT_PASTE)

It's not possible to adapt the sub-commands of the paste command, because
they are dependant on the state of the sheet, and may change based on the
paste location. A simple example is that paste create new col/rows for the
clipboard content to fit the sheet. So there will be ADD_COL_ROW_COMMANDS
in the sub commands in the history, but repeating paste might not need them.
Or they could only needed for the repeated paste, not for the original.

closes odoo/o-spreadsheet#2126

Task: 2508693
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>