[FIX] config: Fix the action version extractor
the tool 'parse_message' was using an incorrect regex to capture the
version of the release. It specifically pose problem with the master
release for which the `alpha` tag was not considered, implying that 2
consecutive master release would have the same tag, which crashes and
ultimately prevent further steps of the github action, like the
publication on NPM.

We can rather rely on the information of the package.json file since
it's properly updated.

closes odoo/o-spreadsheet#2124

X-original-commit: d2361e21c3f01d4de5f7183f7ba00ae236175e24
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>