[IMP] config: tag npm publications
We currently do not specify any tag when publishing on NPM, which
defaults to `latest`. Since we don't want old versions to be tagged as
such, we add a tag in `package.json` to control which version has the
flag `latest`.

Note that according to the NPM documentation[1], the tag cannot match a
semVer range [2][3] which is why we use a special tag that only
indentifies the major version of the package.

[1] https://docs.npmjs.com/cli/v9/commands/npm-dist-tag#caveats
[2] https://semver.org/
[3] https://www.telerik.com/blogs/the-mystical-magical-semver-ranges-used-by-npm-bower

Part-of: odoo/o-spreadsheet#2136