[FIX] popover: adjusted z-index for popover and grid composer
Currently error tooltip is overlapping composer assistant when one tries to
modify arguments within the grid composer.  It arises because of a conflict in
the z-index values between the popover and grid composer components.

To resolve this problem, the z-index of the popover component was reduced.
Now, the popover is behind the grid composer and allows user to view composer
assistant properly .

task-3211697

closes odoo/o-spreadsheet#2205

X-original-commit: a35eeba4947671f0855175356f08ee8e2c6be20b
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>