[FIX] range: show ref error after removing col/row
Previouly after we remove cols/rows that are used in formulas, the
formula will not show error reminding invalid references, but keep the
previous range and give the wrong results.
This commit fixes this problem, so after deleting rows/cols the formulas
refer to removed ones will show errors.

The tests are co-authored with Adrien Minne (adrm).

task 2719611

closes odoo/o-spreadsheet#2215

X-original-commit: 9189b71be37417a6df5a11dca802ead6854c1738
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>
Signed-off-by: Yang Chenyun (chya) <chya@odoo.com>
Co-authored-by: Adrien Minne <adrm@odoo.com>