[FIX] sheetview: correctly auto-size spreadsheet
TLDR default sheet view size is now 0 instead of 1000

Sizing o-spreadsheet has always been difficult because it auto-size itself.

We've had cases where it would start with a sensible size but then shrink to
0 height (ever seen that when debugging QUnit tests ?).

We rely on o-grid-overlay's natural size in the DOM to actually sync the size
in the model.
However, its natural size was influenced by the initial state of the model.
Because the default size in the model was (1000, 1000) the canvas is sized
to (1000,1000) and since the canvas is a child of o-grid-overlay, its natural
size is ~ (1000, 1000).
You had to apply very strict css rules to o-spreadsheet or o-grid-overlay to
force another size.
In particular, it didn't work well if you want css such that "o-spreadsheet
fills the remaining place".
If the available space was smaller than 1000, it couldn't properly "fill the
remaining space" because it was already larger!
A standard monitor height is 1080, taking into account, a header, the topbar
and the bottom bar, you probably never have more than 1000 pixels available.

With this change, it's much easier to size o-spreadsheet the way you want
because it's not influenced by it's initial default size.

closes odoo/o-spreadsheet#2188

Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>