[FIX] viewport: infinite loop with hidden headers
When we move the selection we scroll the viewport so that new selection is
visible. Previously, the strategy to do this was (when moving up):

1) set the viewport offset to the row above the current topmost row
2) adapt the viewport top/bottom to the new offset
3) repeat while the selection is not visible

This doesn't work if the row above the current topmost row is hidden, since
it have the same position has the current row. This led to an infinite loop
with the structure of the code.

With this commit, the hidden rows are skipped in step 1). The loop scrolling
the viewport was also rewritten, so it's impossible to have an
infinite loop even if there is a bug somewhere.

Odoo task 3244431

closes odoo/o-spreadsheet#2351

X-original-commit: 20631704eca24d4044efccf0f73a806442fa3c6b
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>
Signed-off-by: Minne Adrien (adrm) <adrm@odoo.com>