[FIX] ContentEditableHelper: fix getCurrentSelection for Firefox
How to reproduce:
In firefox
- from the grid, input `=sum(`
- hit backspace
-> the selection is set between `s` and `u` while it should be after `m`.

Starting from a cell content `=SUM(`  we have the composer content
```html
<span>=</span>
<span>sum</span>
<span>(</span>
```
When we delete the parenthesis, here is what google does
```html
<span>=</span>
<span>sum{document.selection}</span>
```
and firefox
```html
<span>=</span>
<span>sum</span>{document.selection}
```

The function `ContentEditableHelper.getCurrentSelection` is biased as it would
assume that the selection would always end up in the deepest node (here a node
`#text`).

This commit modifies the function to handle all selection nodes.

closes odoo/o-spreadsheet#2319

Task: 3071026
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>