[FIX] SheetUIPlugin: set column height for multi line content
Previously, when double-clicking on the overlay header to auto-resize
column size, the auto-resizing did not work accurately for multi-line
content. This was due to auto-resizing taking the whole length of the
text without considering whether the content was multi-line or not.

To fix this, we modified the auto-resizing to consider the maximum length
of the broken content for multi-line cells. This will ensure that column height
is properly set for cells with multi-line content.

task_id: 3250821

closes odoo/o-spreadsheet#2497

X-original-commit: 1d7ff54370a3c9a4bec3cc0ac2aeb3c7452ff13b
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>