[IMP] model: core allowDispatch for local commands
Before this commit, local commands were going through the allowDispatch
of core plugins. This wasn't a problem per se, but was wrong from an
architecture standpoint, core plugins shouldn't be aware of what's
happening locally. This also made it so that zones/ranges of local commands
missing a sheetId (because it is implicitly the active sheet) weren't
checked.

Now the local commands don't go through the core allowDispatch, and both
plugins SheetPlugin and SheetUIPlugin are in charge of the generics
allowDispatch.

Task: 3276541
Part-of: odoo/o-spreadsheet#2364