[FIX] lookup: horizontal search in LOOKUP function
Previously, the =LOOKUP function did not work properly when performing a
horizontal search, as it would always pass searchArray[0][index] to the
assertAvailable function parameter.

To resolve this issue, we added a check to determine whether it is a horizontal
or vertical search, and passed the appropriate parameter to the assertAvailable
function accordingly.

closes odoo/o-spreadsheet#2457

Task: 3277145
X-original-commit: 34ef88cbf1bc7e66e3825acb776e5c1019fc4ce5
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>