[IMP] plugin: make CoreView plugins more core
Change the CoreView plugins so they depend as much as possible only of
core commands. This is a first step to have the history in there plugins.

Moved FilterEvaluation and SheetView to ui_stateful plugin. Loaded the
CoreView plugins before the others, since the evaluation result is needed
in FilterEvaluation, and theoretically CoreView plugins should only depend
on core values, so it make sense to load them first. The selection plugin
still need to be loaded first.

This isn't perfect, some  UI commands are still needed in the CoreView
plugins:
- EVALUATE_CELLS
- UNDO/REDO
- ACTIVATE_SHEET (for CF evaluation)

Task: 3288703
Part-of: odoo/o-spreadsheet#2408