[REV] test: make test tsconfig work in vscode
This reverts commit c47fcd7a445ba51c3d20fba2cdf08a28731c44af.

Here goes the story:
we asked to upgrade nodejs version on runbot (which was version 12)
It was upgraded to version 20 in the global master Docker image.
It worked fine for odoo...but not for o-spreadsheet.
The build hangs indefinitely at the end of the tests.
We can't reproduce locally, even when using the same Dockerfile.

We found out this commit was the culprit, but we don't know why...

closes odoo/o-spreadsheet#2419

Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>