[IMP] model: dispatch initial revision to core only
Currently, the initial revisions is dispatched to all plugins, core
and UI.

Several issues with this:

- Any revision with an UPDATE_CELL command would trigger an evaluation.
This means we evaluation many many times, and we don't need to (only
the very last matters). This really can be a perf killer when loading
a spreadsheet.

- if an initial revision adds a column before column A, the initial
selection (supposed to be A1) is shifted to B1.

With this commit, initial revisions are dispatch before setting up
UI plugins.

closes odoo/o-spreadsheet#2421

Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>