[IMP] spreadsheet: add locale handling
Add the handling of locale in spreadsheet, which will impact the formats
of numbers and dates. The locale is changed from the `SettingsPlugin` and
its corresponding side panel, and is shared for all users.

Most of the format helpers and the helpers to convert numbers to/from string
now take the locale as argument to adapt their result based on the user
locale.

The tokenizer also is localized now, because for some locales (eg. French)
we need to change the argument separator from `,` to `;` for it to not be
the same as the decimal separator.

closes odoo/o-spreadsheet#2483

Task: 3222401
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>