[IMP] tests: use keyUp/keyDown helper everywhere
The helper shorten the tests and ensure that all the keyboard event bubbles
and are cancelable. The helper is async, and can be awaited to wait for
the next render if needed.

The event targets the active element to try to be close to the reality.
Helpers `focusAndKeyUp/Down` were also created for the tests where we need
to focus an element manually.

Task: 3302326
Part-of: odoo/o-spreadsheet#2080