[FIX] viewport: warning message on resized viewport size
Fix warning not closing issue caused by changing the value of
DEFAULT_SHEETVIEW_SIZE to 0 from 1000 in a previous PR #2188. The problem
arose due to xRatio and yRatio becoming NaN and Infinity prior to mounting.

To resolve this issue, Added an if condition that returns the function if either
xRatio or yRatio is Infinity.

Task-3318806

closes odoo/o-spreadsheet#2504

X-original-commit: 254ac6c284eda900c0126be8b932ce6a4944cf83
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>