[FIX] conditional_formatting: fix cf type checkboxes
The checkboxes of the CF side panel weren't working properly if we clicked
right between the checkbox input and it's label. The state of the component
changed, but the checkbox wasn't checked.

This is because we used `t-attf-checked` to defined the state of the checkbox,
which returns a string instead of a boolean. Changed it to `t-att-checked`.

closes odoo/o-spreadsheet#2530

X-original-commit: eaa83fcb5a855175452155e1ce0fa85247271ab8
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>
Signed-off-by: Minne Adrien (adrm) <adrm@odoo.com>