[REF] Selection: Let caller control selection visibilty
The first design of the processor was too intertwined with the viewport,
and we ended up designing some commands mode/types  so that they would
simultaneously:
- fit our need to distinguish use cases that required the viewport(1)
  to move or not.
- still made sense from a functional point of view (the types & modes
  reflect the intention)

Unfortunately, we already have 9 different combinations of types, only
two third of them really used in order to distinguish intensions) and
a bug was reported (see #2369) taht ipplies that this does not cover
all the expected behaviours.

We might have mistaken the role of the processor by thinking the API
so that the viewport would determine itself how to react to an update
of the selection when we actually implemented are methods that suit the
caller's will to reflect the change of the selection in the viewport or
not.

The processor is just a channel that should:

- dispatch the selection changes to the right listener
- in some cases notify the viewport that it should react to that change

This commit gets rid of the `SelectionEvent.type` in favor of specific
options to explicitely request some side effect behaviours like:
- scroll the viewport to display the newly selected cells
- whether the zone should be treated as an unbounded zone or not

closes odoo/o-spreadsheet#2553

Task: 3348042
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>