[REF] composer, grid: homogenize event handling
In the composer, we had different strategies to handle events. To handle
`NumpadDecimal`, we intercepted the event and did the modification to the
composer content by hand.

For new line event, we intercepted Ctrl+Enter, then dispatched an InputEvent
`insertParagraph`, then we still had to handle the event by hand. This
commit remove the step of dispatching a new InputEvent, which was useless.

Note on the why: the `insertParagraph` should have been handled by the
contentEditable, but isn't since it's dispatched by the code, and is not
trusted. Thus it needed to still be handled by hand.

This commit also made use the same helper in both `grid.ts` and `composer.ts`
to transform the keyboard event into strings. We now keep the native
JS CamelCase for events keys instead of upper-casing everything.

closes odoo/o-spreadsheet#2686

Task: 3380401
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>