[FIX] SortPlugin: fix crash whith empty-ish cells
**Current behavior before PR:**
When sorting the cells (in `sortZone` method), it is editing the cells
directly inside the loop. When it dispatches the "UPDATE_CELL"
command, the content of the cell can be considered as empty and the cell
will be removed as well as the corresponding `cellId` from the
`SheetPlugin.cellPosition` array.
Therefore, it won't be able to retrieve the position of the cell in a
future iteration of the loop to retrieve its value (with
`this.getters.getCellPosition`)

**Description of the fix:**
Store the cell updates in an array to apply them all at once at the end.

opw-3422772

closes odoo/o-spreadsheet#2702

X-original-commit: b829925f75eb3d51455a782404e60d982d8c9b99
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>