[FIX] Find&Replace: shorten the debounce on find search
The debounce of 400 ms made the search feature feel sluggish.
This becomes especially weird since we fixed the look of the disabled
buttons when a result in not found. The button is disabled by default,
hence will stay that way after typing a character and the first
debounce is executed.

closes odoo/o-spreadsheet#2699

Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>