[FIX] find_and_replace: disabled button css and shorten the debounce
Previously  find and replace side panel buttons did not have
the proper style when they were disabled.

This commit will fix that CSS issue.

The debounce of 400 ms made the search feature feel sluggish.
This becomes especially weird since we fixed the look of the disabled
buttons when a result in not found. The button is disabled by default,
hence will stay that way after typing a character and the first
debounce is executed.

closes odoo/o-spreadsheet#2727

Task: 3432224
X-original-commit: 8ad05356d630b0ced907bf6d876b776bd3565bd3
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>
Co-authored-by: Rémi Rahir <rar@odoo.com>