[IMP] evaluation: allow formulas to handle error
The purpose of this commit is to allow the formulas
to return errors as results.

Functionally we want formulas like TRANSPOSE to be
able to transpose an error without invalidating the
entire evaluation process.

We also want formulas like LOOKUP not to be interrupted
when the search range has one or more errors.

To make this change we accepted errors in input and
output on the formulas,

It nevertheless remains possible to throw errors in the
formulas:
it is no longer the evaluation process at the cell level
which is stopped but the evaluation process at the formula
level.

Technically there is no longer a "try catch" at the
evaluator level,
here each formula is encapsulated in a “try catch”
when it enters the formula register

Other point, the _lastFnCalled present in the evaluation
context and allowing to know which is the last formula to
have been called before the raise  of an error is no longer
used. The information is available during the encapsulation
of the formula in the "try catch".
So, _lastFnCalled has therefore been removed from the
evaluation context.

Task: 2969401
X-original-commit: 78e377f4d72881076a31bac10c1994e0cc8305c1
Part-of: odoo/o-spreadsheet#3454
Co-authored-by: Lucas Lefèvre (lul) <lul@odoo.com>