[IMP] package: convert to ES module
This commit converts the package to an ES module.

It was a CommonJS package (with `require(...)` imports) which is
the default for node.
ES module is the modern approach.

Existing CommonJS files are renamed with the exention `.cjs` to make it
explicit and to tell node to load them as CommonJS files.
Those files could be converted to ES syntax in the future.

The goal is also to upgrade rollup to version 3 which forces a strict ES module
config file by default (currently, I'm not sure how it works. It is written as
an ES module, but loaded by node as a CommonJS file ? rollup 3 doesn't work
the same way)

Part-of: odoo/o-spreadsheet#2808