[IMP] Spreadsheet: prevent change of props model
While a model is passed as a props to the  `Spreadsheet` component,
it is used in the setup of the latter to register bus events, build a
subEnv and potentially other steps that bind the component specifically
to that instance of `Model`. It will take some thinking to properly
address that issue but in the meantime, we stop the update of the props
`model` to alert the developpers that would try to integrate the
ligrary.

Part-of: odoo/o-spreadsheet#2811