[REF] functions: simplify arguments types using type Maybe<T>
In this commit, we introduce a new generic type called `Maybe<T>`, which
signifies the potential absence of a value.
This pattern is widely used in various programming languages, such as Haskell,
Elm, F#, Rust or Scala.
While F#, Rust and Scala implement it under the name `Option<T>`, the concept
is the same.

This brings the opportunity to eliminate two existing types:
`PrimitiveArgValue` and `PrimitiveArg`. These types lack clarity in their
naming and often require jumping to their definitions for proper understanding.

With this straightforward concept of `Maybe`, the revised version simplifies
the understanding of types.
Now, the code only deals with two distinct types, namely `CellValue` and
`ValueAndFormat`, instead of managing four separate types:
`CellValue`, `ValueAndFormat`, `PrimitiveArgValue`, and `PrimitiveArg`.

closes odoo/o-spreadsheet#2812

Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>