[FIX] demo: console error messages
Error in the Demo component are caught and print in the console. Most
of the errors are caught by Owl and wrapped, in which case we want to
print the `error.cause` property. But some errors aren't caught by Owl
and thus `error.cause` is undefined (eg. non-defined component in template).

Also change the errors in Demo to be notifications instead of blocking
`window.alert`. Having blocking errors is annoying when developing since
it totally blocks auto-refresh and we need to manually close the popup
to continue working.

closes odoo/o-spreadsheet#2813

Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>