[FIX] GridOverlay: Hovering over a figure should not open popovers
Previously, when hovering over the chart, the cell popover would erroneously open.
This issue occurred because updateMousePosition assigned the clientX and clientY
values, and getPosition found the cell's position, creating a bug when hovering
over the chart.

To address this problem, a conditional check has been added to ensure that
`gridRef` is equal to the event target.

closes odoo/o-spreadsheet#2877

Task: 3475001
X-original-commit: def91b6ca7e34d9ef889ae6a8d06faf048245b0d
Signed-off-by: Adrien Minne (adrm) <adrm@odoo.com>