[REV] components: prevent mouse event defaults on drag-and-drop
Preventing the default events in DnD unfortunately introduced a 'too generic'
behaviour. Due to the current state of the model and component synergy,
we have some limitations when drag-and-dropping.

Specifically, an issue can arise when a user would input a character
just after starting drag-n-drop (i.e. hit a key just after a mousedown
on the grid). In this situation, inputting the character will make a
switch to a state of cell edition, which is not compatible with the
callbacks provided to the drag-n-drop. So any further mouve of the mouse
will then invoke the SelectionStreamProcessor, that will contact the
`EditionPlugin` but since we are not selecting a cell, it does not make
sense and is simply ignored but the plugin[1]. The user will end up in a
situation were they think they are selecting a zone but nothing happens.

Commit 1178bbec was introduced to fix a very specific issue but doing
so, removed the ability to target another DOM node on `mousedown` event,
which happens to be very helpful to lose the focus of the `hiddenInput`
focused by in grid when we click somewhere on `GridOverlay`.
It turns out the original addressed issue can be handled case-by-case by
preventing the default behaviour of the mousemove callback. This
revision suggests to take a more surgical approach and prenvent the
default behaviour where it is required (uncalled-for pre-optimisation is
bad, m'kay?). If we were to require more generic limitations in the
future, then we will need to introduce a way to invalidate drag-and-drop
under given conditions.

[1] in later versions (16.5-alpha and onwards) it's no longer ignored but will
rather trigger an error...

closes odoo/o-spreadsheet#2828

X-original-commit: a3b2e7cb67d44b052f38da6e2453919ea2b5148e
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>