[IMP] export: do not export defaulf style values
Open the background color picker and click on
"No Color".
Double click on the Bold icon (to set the bold and remove it)

Now export the spreadsheet.

You'll see something like
```
{
  ...
     "cells": {
       "A1": { "style": 1 },
     },
  ...
  "styles": {
    "1": {
      "fillColor": "",
      "bold": false,
    }
  }
}
```

The style is actually useless.
If many cells have this useless style, it can quickly bloat the
export with lots and lots of useless lines (which can end up being
committed in odoo's source code of dashboards).

With this commit, default values are no longer exported.

(also removed an unused constant)

closes odoo/o-spreadsheet#2876

Task: /
Signed-off-by: Adrien Minne (adrm) <adrm@odoo.com>