[REF] Scorecard are now drawn on a canvas
While we are now able (from previous commit) to export the gauche
chart as an image, the scorecards were still non exportable as they
were directly rendered as html element and not on a canvas. This
commit change the way we are drawing scorecard chart, by using now
a canvas. The drawing logic has been extracted from the scorecard
element to be able to use it also during the export process.