[FIX] evaluation: compute cell only once
At the first evaluation, cells are computed in the order they are defined.
When a formula references another cell, the cell is computed (if it's not
already computed). However, the result is currently not stored. Therefore,
if another formula references the same cell, it will be recomputed!

It can drastically slow down the evaluation as the same cells can be
recomputed an enormous amount of time.

In a (quite big) spreadsheet prepared for a demo at OXP, evaluation goes from
more than 30 seconds to less than 500ms

Part-of: odoo/o-spreadsheet#2971