[FIX] evaluation: don't build formula content for errors
Building the string content of a string is not free for
formulas with dependencies (it's built dynamically, based on
the ranges.)

With this commit, we omit building it if the cell is resolved
to an error. The content was ununsed anyway.

The total time spent buliding the content in `handleError`
could go as high as 6% of the loading time of a spreadsheet
with many pivot functions.

closes odoo/o-spreadsheet#2980

Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>