[IMP] headers: add header position UI plugin
This commit add the `HeaderPositionsUIPlugin` which computes and stores
the positions of the headers.

Before this the headers positions weren't stored, and needed to be
computed by adding together the size of all the preceding headers in a
loop. This was a performance issue when scrolling on larger sheets.

closes odoo/o-spreadsheet#3049

Task: 3563624
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>