[FIX] popover: fix popover position when too much items
Task Description

When opening a topbar menu with too many items in it (which can
happen, mainly, in Odoo with too many data sources), the menu is
not correctly placed. This comes from the fact that, in the
computation of the position, we check if the popover should render
from the bottom using the original children size instead of the
real size of the popover (the min between the 'maxSize' that the
popover can be and the children size).

Related Task

closes odoo/o-spreadsheet#3091

Task: 3418671
X-original-commit: 7a6bcf074acc90e9f457526ecd923179d0760508
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>