[REF] tests: dispatch both input and change events
Currently in the test, we often only dispatch either an input or a change
event. But in practice, both events are fired when modifying an
input. This could lead to inconsistencies between the test and the
reality.

This commit changes the setInputValueAndTrigger to dispatch both events.
It also make the helper async with a nextTick at the end to shorten the
tests.

closes odoo/o-spreadsheet#3057

Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>