[IMP] actions: don't create actions at each rending
With this commit, we are no longer creating the full action at every
rendering.

`createAction` has a hidden cost of generating a uuid if the action does not
have one.
When scrolling really fast (to trigger lots of rendering), `getRandomValues`
was taking almost 4% of the compute time. Now it's no longer called at every
rendering.

Part-of: odoo/o-spreadsheet#3086