[IMP] range: order only if needed
Every range created was re-ordered, which actually recreates a new range, even
if the range was already ordered.
Now, we do not re-creates the range if it's already ordered.

The gain is hardly measurable (it's less than ~40ms over 10+s when loading
a sheet with 520k cells)

Part-of: odoo/o-spreadsheet#3086