[REV] Charts: Do not rely on history for the chartRuntimes
Since [1], the commands "UNDO" and "REDO" no longer invalidate the chart
evaluation as it was supposedly handled by the plugin history. However,
the ChartEvaluation plugin would lazy create the chart runtime and
update the history outside of the dispatch context (the changes are
therefore not recorded).
After some consideration, we consider that the storage cost of the
objects `ChartRuntime` is greater than having to invalidate/recompute it
at each UNDO/REDO. We therefore revert the change.

[1] https://github.com/odoo/o-spreadsheet/pull/2408

closes odoo/o-spreadsheet#3149

Task: 3578417
X-original-commit: 8b5877679c7d6c81e4d6035180e7efb2471530a2
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>