[FIX] F&R: Fix css rule
The padding rule did not respect the web standards. It just happened to
work on our demo sheet because the page is run in quirks mode, a
mode that support web pages prior to the instoduction of the web
standards. This rule probably never worked when the component was
embedded in a "modern" web page.

closes odoo/o-spreadsheet#3140

Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>