[FIX] lookup: cache string normalization
This commit caches string normalization used in lookup functions.
`normalizeString` is called for every element of lookup ranges. Those
ranges can be large.

In the production spreadsheet mentionned on the ticket, it was taking more
than 13% of the loading time. Now it's 0.0%

`normalizeString` was called 19M times while it has only 664 different
input values.

opw-3588401

Task: 3588401
X-original-commit: 943a7e7f843f5aef15b06c91a0609408cebc0e87
Part-of: odoo/o-spreadsheet#3224