[IMP] highlight: draw highlight over cell borders
Before this, because of our computations and canvas' shenanigans, the
highlights were not drawn exactly over the cell borders, but slightly
beside it. This was particularly visible when drawing only the borders
of the highlight without fill color, and wasn't very pretty.

This commit made use of the same computations to draw the highlights
than the computations of the cell borders, which makes the highlights
be drawn exactly over the cell borders.

Task: 3571796
Part-of: odoo/o-spreadsheet#3304