[FIX] evaluation: cache range references values
When a function references a range `=SUM(A1:B10)`, a new 2d array is rebuilt
with the cell values for every range in every functions.

However, for LOOKUP functions, it's very likely to have many cells with a
LOOKUP function looking for a match in the exact same range.

In the customer spreadsheet, `range` self-time goes from ~15% to almost nothing
(0.2%). There's another spreadsheet where many cells references the same range
with 100k cells. Loading the spreadsheet was taking 26s before and 7s after.

opw-3588401

closes odoo/o-spreadsheet#3210

Task: 3588401
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>