[FIX] Autofill: Automatic autofill based on content
The Automatic autofill is supposed to consider the cells that are
seemingly empty be still have some content (for instance, a formula resulting in
an empty string). The behaviour was broken when we refactored the evaluation
in `saas-16.1`.

Note that the current fix will no longer be correct once the spreading formulas
will be introduced.

closes odoo/o-spreadsheet#3214

Task: 3592859
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>