[FIX] Spreadsheet: Allow shift-scroll to work on mac os
On mac os, using the wheel of a mouse while pressing the `shiftKey`of
the keyboard will automatically swap the event deltaX and deltaY[1].

This revision proposes to let MacOs do the swapping so we can have the
same behaviour cross-platform by scrolling while pressing the `shift` key.
This approach is consistent with the experience with both Excel Online
and Google Sheets.

[1] Note that this only applies to external mouses. The trackpad of a
macbook will behave 'windows-like' but it can track in both directions
by default.

closes odoo/o-spreadsheet#3229

Task: 3603771
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>