[FIX] icons: use correct comment syntax in xml files
The comment syntax in xml files is <!-- comment --> and not /** comment */.
This could lead to errors when the xml files is parsed, as it's read as
a text node. It works right now because the xml is parsed entirely, and
owl filter nodes that has attribute `t-name`.

But in the case it's parsed partially (on demand for example), it leads
to errors.

Part of task-id 3601257

closes odoo/o-spreadsheet#3244

X-original-commit: 88a57c6be5e5d20a1ab313f6892651b21786ae7d
Signed-off-by: Vincent Schippefilt (vsc) <vsc@odoo.com>
Signed-off-by: Pierre Rousseau (pro) <pro@odoo.com>