[FIX] SelectionInput: release overriden input
When we replace a selection input by another one with the same id but
while providing initial ranges, the previous input was not automatically
released from the selection which led to a weird state where the
selection processor was captured by an Owner that was no longer
referenced in the plugins/components environment (kind of an orphan
object). This was overlooked in pr #2858.

closes odoo/o-spreadsheet#3290

X-original-commit: 1d1dfbcaf131f51afec3decf45de0b7de16ff5b9
Signed-off-by: Vincent Schippefilt (vsc) <vsc@odoo.com>
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>
Co-authored-by: Adrien Minne <adrm@odoo.com>