[FIX] find&replace: Only switch sheet from the side panel
The recent improvement in [1] introduced a faulty behaviour, which is to
try to dispatch commands in the finalize step of a dispatch. The
functional intension was to let a user interact with the sheet and that
such actions would automatically be mirrored in the f&r sidepanel.
Ultimately, this would prove problematic as we need to perform the
search in the finalize (because of the evaluation process) and that
could lead to a tentative to switch sheets in order to let the user jump
at the position of the match.

This revision addresses the issue by modifying the functional behaviour.
As of now, if the currently selected match is not on the current sheet,
the user will be redirected to the match position IIF they dispatch a
"find&replace" related command (i.e. if they interact with the F&R
sidepanel).

[1] https://github.com/odoo/o-spreadsheet/pull/2330

closes odoo/o-spreadsheet#3312

Task: 3618775
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>