[IMP] demo: don't double crash on launch error
When an error occurs on launch, the demo would crash twice: once when
the error is raised, and once when we tried to use `notifyUser` to
display the error message. `notifyUser` tried to add a notification to
the `o-spreadsheet` element, which was crashing since it wasn't mounted.

This commit fixes the issue by displaying notification on the document
body rather than on the `o-spreadsheet` element.

closes odoo/o-spreadsheet#3335

X-original-commit: 63af9b38e606386fa65ba7e061221ed10835baf0
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>
Signed-off-by: Adrien Minne (adrm) <adrm@odoo.com>