[REF] SelectionInput: rewrite plugin with store
This commit refactors the SelectionInput component's state and business logic.
Previously, the state of an individual SelectionInput was handled through a
plugin-like structure, overseen by the "selection input manager" plugin.
This manager was responsible for coordinating interactions between
SelectionInput components, primarily focusing on which one is currently active.

However, this approach introduced a significant amount of boilerplate code.
The manager functioned as a proxy between the component and the individual
plugin-like structure. There was also a requirement for setup and cleanup
commands for each component upon mounting and unmounting, leading to the
necessity for two renderings before a correct and consistent setup of both the
component state and the "plugin."

Now, each SelectionInput now has its own local store, and each component
communicates exclusively with its own store.
To manage interactions between these stores, a straightforward "FocusStore"
has been introduced (the name is subject to change as I don't really like it).

The old "manager" has been entirely removed.

Part-of: odoo/o-spreadsheet#3304