[FIX] sheet_interactive: rename sheet in readonly mode
The commit rectifies an issue where users in readonly mode were able to invoke
the sheet renaming dialog by double-clicking on its name. However, they were
unable to complete the renaming process in readonly mode.

This change amends the behavior by implementing an early return, preventing the
invocation of the renaming dialog altogether when the application is in
readonly mode.

Task:: 3621086

closes odoo/o-spreadsheet#3345

Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>