[IMP] side_panel: add basic components
This commits introduce two new components aimed at being used in the
side panel. The first one is a Checkbox, the second one is a Section.

The Checkbox is a simple checkbox, with a label and a value.
The Section is a container for other components, with a title.

These components will help to build side panels with a coherent look
and feel for all the side panels.

Task: 3635897
Part-of: odoo/o-spreadsheet#3355