[IMP] PieChart: exclude negative values
Negative values do not make sense in a Pie chart as each data is
supposed to represent the part of a whole. While chart js supports
negative values by handling their absolute value, we consider that it's
more misleading than anything. This revision filters out the negative
values of the datasets.

closes odoo/o-spreadsheet#2972

Task: 3510971
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>