[IMP] formula: add INDIRECT function
Task Description

This task aims to add the INDIRECT function in o-spreadsheet. This
has been implemented by returning the value of the referenced cell
which means that we are still not able to do something like
`=SUM(INDIRECT("A1"):INDIRECT("B3"))`, but other uses of this
methods should be available.

Related Task

Task: 3203719