[REV][FIX] side_panel: cf rule range bugs
This reverts commit 4a956f3eafe299badcc9479dc1eecd7f55248b25.

the task was not fixing bugs (in the sens of tracebacks or corrupted
behaviour) but rather seemingly backporting a feture introduced in
later versions. Since the implementation was not accounting for every
case and that it' not fixing any behaviour, just improving the user
experience, it's better to  cancel it.

closes odoo/o-spreadsheet#3450

X-original-commit: ce7412e020f2cdbf3df3c1c92e26c90686a4984b
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>