[REV][FIX] side_panel: cf rule range bugs
This reverts commit f7525195cb535ed673133b34eac7aad1fc6840cc.
The task was not fixing bugs (in the sens of tracebacks or corrupted
behaviour) but rather seemingly backporting a feture introduced in
later versions. Since the implementation was not accounting for every
case and that it' not fixing any behaviour, just improving the user
experience, it's better to cancel it.

closes odoo/o-spreadsheet#3452

X-original-commit: 1c77cc39102bb4af53e2a543b4c60325f8ea3526
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>