[IMP] tokenizer: add args AST to function context
This commit adds more information to the function context of tokens which are
part of a function.

The goal is to easily identify each function arguments. It can allow to
implement custom auto-complete based on the function arguments.

As an example, given the custom function we have in Odoo
`=ODOO.PIVOT(1,"expected_revenue","create_date:month","01/2024","stage_id",4)`

If we want to provide an auto-complete for field names or group values,
we must know the pivot id, which is the first argument.

Because an argument can be composed of several tokens (`SUM(1+2)`) and
the business code of the auto-complete probably wants a simplified version
of the arguments (don't want to handle space tokens for instance), we
provide the AST of each argument. If the argument is not a valid expression,
it's letf undefined.

Task: 3662965
Part-of: odoo/o-spreadsheet#3401