[IMP] table: introduce table style
This commit improves the old concept of tables (filterTable) that only
allowed to filter the table. Now, the table can be configured and
styled in a side panel.

Introduce the TableStylePlugin that compute and store the style of
each cell of a table. The cellTableStyle is then combined with the
cell style and the CF style in ui_sheet.

The only styles available for now are the styles defined in the
TABLE_PRESET constants. These are based on the OpenXML specification
of table styles, with some color changes.

Further improvements might include:
 - allow to define custom table styles
 - dynamic tables that take the size of a spread formula result

closes odoo/o-spreadsheet#3536

Task: 3636461
Signed-off-by: Vincent Schippefilt (vsc) <vsc@odoo.com>