[FIX] evaluation: allow empty error message
Currently, a function cannot throw an error with an empty error message.
(e.g. `new EvaluationError("")`) because of the `||` fallback.

An error without message behaves just like an error but won't render
the small red triangle on the top right corner of the cell and won't
show the tooltip with the error message.

The "Loading..." error in Odoo is such an error.

closes odoo/o-spreadsheet#3535

Task: /
X-original-commit: 4b7ff5f491ac0da96db986da809c7579926d38bc
Signed-off-by: Alexis Lacroix (laa) <laa@odoo.com>
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>