[IMP] array formula: `getSpreadPositionsOf` now returns root position
Before this the getter `getSpreadPositionsOf` only returned the
position of the cells an array formula was spread to, minus the
cell the array formula was in. This made it impossible to differentiate
between a cell with a normal formula, and a cell with an array formula
returning a 1x1 matrix.

Now the getter returns the root position of the array formula as well.

The getter `getArrayFormulaSpreadingOn` was also modified to work when
calling it on the root position of an array formula.

There was also a bug where the spread relation weren't properly cleaned.

Task: 3707037
Part-of: odoo/o-spreadsheet#3624