[FIX] Selection: Loop selection based on content
Similarly to the issue addressed in PR #3214, we want `loopSelection`
to consider seemingly empty cells when considering the expansion zone.

This revision splits the behaviour of the selection processor for
`loopSelection` and the other navigation features (like cluster
jumping) as we still want to ignore cells with an empty evaluation in
those cases. The reason being that the user could be lost if, when
navigating with their keyboards, they end up with their selection on
empty cell.

As for #3214, the fix will need to be adapted to account for the
spreaded formulas.

closes odoo/o-spreadsheet#3663

Task: 3709340
X-original-commit: 2a21fea2dbaf4f792ef19c797b1b66ebe024c684
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>