[FIX] Selection: Loop selection based on spreaded formula
Similarly to the issue addressed in PR #3214, we want `loopSelection`
to consider seemingly empty cells when considering the expansion zone.

This revision splits the behaviour of the selection processor for
`loopSelection` and the other navigation features (like cluster
jumping) as we still want to ignore cells with an empty evaluation in
those cases. The reason being that the user could be lost if, when
navigating with their keyboards, they end up with their selection on
empty cell.

This PR is the adaptation of #3573 to support cells that contain the
result of a spreaded formula.

closes odoo/o-spreadsheet#3664

Task: 3709340
X-original-commit: 2e69e1be48867c24b95b6a4c860323fbe604f55b
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>