[FIX] SelectionInputsManagerPlugin: bad ranges break selection input
In the past, entering invalid ranges (e.g., 'TEST') in the selection
input, hitting enter, and subsequently attempting to select ranges from
the grid would result in a malfunction.

This commit addresses the problem by eliminating the if statement
responsible for validating ranges, as it was identified as the root
cause of the issue.

Task ID: 3626171

closes odoo/o-spreadsheet#3724

Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>