[REF] Renderer: use layer name for rendering
Before this commit, the layer numbers were used in the different
renderers. This means that when the lib is compiled an used in
javascript, it was impossible to know on which layer you were writing.

After this commit, we are using the layer names in the callback of
`drawLayer` to be more readable.

It is now also possible to add a new rendering layer from outside
o-spreadsheet using `AddRenderingLayer` and specifying its priority

closes odoo/o-spreadsheet#3603

Task: 3719197
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>