[IMP] chart: Improve scorecard design
The aim of this task is to improve the design of the scorecards to
take advangate of the available size instead of having some strange
rule for the fontsize of some parts. The key value now take as much
place as possible, and the baseline description is able to split into
two lines to make the baseline value bigger.

Related Task

Task: 3716974
Part-of: odoo/o-spreadsheet#3650