[IMP] Find&Replace: support inverse iteration on search results
At the moment, the user can press `Enter` in the search input to iterate
through results. In most serach related features, pressing shift inverts
the iteration order.

We already support this behaviour with the `Enter` and `Tab` shortcuts
when navigating through the grid.

This commit applies the same logid to the find&replace search input.

closes odoo/o-spreadsheet#3651

Task: 3736442
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>