[IMP] cell: avoid useless UPDATE_CELL
Task Description

This task aims to avoid the dispatch of an UPDATE_CELL command when
the cell to update is already in the same state that the one we want
to get with that UPDATE_CELL command. This is done by adding an
allowDispatch to make sure there is a difference between the cell
state and the command payload.

Related Task

closes odoo/o-spreadsheet#3708

Task: 3716689
X-original-commit: 5a8ec41267f5b3847861d3fa7cc0a79e4786f3fb
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>
Signed-off-by: Hendrickx Anthony (anhe) <anhe@odoo.com>