[FIX] FigureContainer: no selection of a figure content
Currently, when playing with the resize corners of a figure, you can end
up selecting the inside of the figure, which gives it a highlight. This
has no functional purpose and the only way to get rid of it is to select
another dom node than a figure (like the grid or the composer).

closes odoo/o-spreadsheet#3722

Task: 3752290
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>