[REF] filter: move filters icons inside grid overlay
This commit moves the filters icon overlay from `grid.xml/dashboard.xml`
to `grid_overlay.xml`. This has several benefits:

- since the icons are positioned inside the grid overlay, they don't
have to worry about the grid/dashboard position offset. The GridOverlay
handles that.
- we can call the onMouseDown of grid from the FilterIconOverlay,
selecting the cell with all its side effect (eg. closing the current
grid composer on left click, opening the context menu on the correct
cell on right click, ...)

The annoying thing is that we need to prevent the mouseDown to close
the filter menu when clicking on the filter icon. This is handled
on the click of the filter icon. Otherwise, the menu toggle won't work.

Task: 3754487