[FIX] evaluation: accept 1x1 result array in sub-formula
Before this commit, when a formula function
expected a simple argument, and it received
an array, we transformed the array into a
simple argument if the array was of size 1x1

However, this process was only done when
reading the references to the ranges.

This commit extends the process when the result
array comes from a sub-formula and no longer
range references.

Task 3756474