[IMP] model: expose canDispatch to plugins
Core commands don't go through allowDispatch if they are dispatched
from a core plugin, because we assume the core plugins know what
they are doing. But sometimes it'd be useful to get the result of
canDispatch for a core command, so we don't dispatch invalid commands.

It can also sometime be useful for a command to call `canDispatch`
on a known sub-command in its allowDispatch method.

closes odoo/o-spreadsheet#3744

Task: 3709540
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>