[FIX] Migration: Propagate datafilter correction on every version
The fix proposed in https://github.com/odoo/o-spreadsheet/pull/3632 will
only work if the data are actually migrated, i.e. when the loaded data
is from a version before 12.5. Unfortunately, a lot of time passed
before we discovered the issue and the spreadsheet version was updated
several times.

To ensure the data integrity accross all versions, we have to duplicate
the migration of the data filter so that every current active version
can benefit from it.

closes odoo/o-spreadsheet#3761

Task: 3728009
Signed-off-by: Pierre Rousseau (pro) <pro@odoo.com>
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>