[FIX] headers_overlay: unhide cols/rows icon position broken
Previously, there was an issue with the position of the unhide cols/rows
icon, which has now been rectified. This fix involves setting the icon's
position to absolute, applying bootstrap classes, and removing CSS from
the ts file for o-unhide-buttons.

Furthermore, when the hiddenItems consist of the first or last column/row,
bootstrap classes are now utilized to appropriately hide one of the icons.

Task ID: 3696210

closes odoo/o-spreadsheet#3905

X-original-commit: 97d200f6ecc0b402fe54c7dd37c63bc606e09a95
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>