[IMP] table: add table colors to custom colors
When inserting a table with a style in the sheet, it could be useful
for the user to be able to paint cells in the same colors as the table
cells.

This commit adds the table colors to the custom colors plugin, so they
are available to the user.

Another idea would have been to replace the color pickers colors by
the table colors. But then we end up with dull and sad colors in the
color picker so we gave up on that.

Also fixed some bugs and strange behaviors in the custom colors plugin:
- we compute and store the custom colors at each finalize, but the
getter `getCustomColors` was re-computing the colors instead of using
the stored ones.
- the colors stored by the plugin weren't loaded at import until a
command modifying the colors was executed.
- the colors were filtered/validated multiple times (once when computing
the colors, and once when adding them to the plugin state).
- the colors were sorted, then added to a POJO, making the sorting useless.

Task: 3775936
Part-of: odoo/o-spreadsheet#3776