[FIX] FiguresContainer: chart resizing broken
Previously, the chart resizing didn't work properly when the viewport
was scrolled and only part of the figure was visible.

This commit addresses the issue by enhancing the condition to check if
the figure extends beyond the header boundaries. Now, this condition
also considers scrolling, ensuring proper chart resizing.

Task ID: 3752060

closes odoo/o-spreadsheet#3785

Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>