[FIX] FiguresContainer: chart resizing broken
Previously, the chart resizing didn't work properly when the viewport
was scrolled and only part of the figure was visible.

This commit addresses the issue by enhancing the condition to check if
the figure extends beyond the header boundaries. Now, this condition
also considers scrolling, ensuring proper chart resizing.

Task ID: 3752060

closes odoo/o-spreadsheet#3829

X-original-commit: e87e56c8b9de0b3cdd5eaa4586a4f8a643f3db67
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>
Signed-off-by: Dhrutik Patel (dhrp) <dhrp@odoo.com>