[FIX] FiguresContainer: chart resizing broken
Previously, the chart resizing didn't work properly when the viewport
was scrolled and only part of the figure was visible.

This commit addresses the issue by enhancing the condition to check if
the figure extends beyond the header boundaries. Now, this condition
also considers scrolling, ensuring proper chart resizing.

Task ID: 3752060

closes odoo/o-spreadsheet#3822

X-original-commit: d503e6d4b15719fbda1a6e3833380a0b9119526a
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>