[IMP] tables: allow creation of custom table style
This commit introduces the ability for the user to create custom
table styles. The table styles are now managed in a new `TableStyle`
plugin instead of being based on the constant `TABLE_PRESETS`. The
old table style plugin was renamed to `ComputedTableStyle`.

The option to create a custom table style is available in the table
style popover, that opens a new side panel to create a table style.

closes odoo/o-spreadsheet#3865

Task: 3802101
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>