[FIX] tests: prevent chaining of `makeTestEnv`
Since the introduction of stores, chaining `makeTestEnv` is wrong and
could cause hard to understand bugs in the tests.

We cannot chain `makeTestEnv` because the first one will create a new
DependencyContainer and inject the model/notification inside it. Then
the second one cannot inject a new model since a model is already in
the DependencyContainer, and cannot create a new DependencyContainer
since it would lose things injected in the first one.

Now chaining `makeTestEnv` will throw an error.

Also fixed a bug in `makeTestEnv` where `toggleSidePanel` was not
properly initialized.

Task: 3802101
Part-of: odoo/o-spreadsheet#3865