[IMP] bottom_bar: prevent selection of hidden sheets
Before activating an hidden sheet had the side effect of making the
sheet visible. But it made possible to activate a hidden sheet in
readonly mode, since the command to un-hide the sheet was prevented.

Now the allowDispatch of `ACTIVATE_SHEET` prevents activating an hidden
sheet. In the bottom bar, we also now  disable the menu item that
unhides & activates an hidden sheet in readonly.

closes odoo/o-spreadsheet#3831

Task: 3801883
Signed-off-by: Vincent Schippefilt (vsc) <vsc@odoo.com>