[FIX] TopBar: Remove hidden functions from the `insert` menu
The introduction of the `hidden` tag of functions in pr #1928 did not
account for the `insert function` menu of the top bar that was
developped at the same time.

This revision ensures that the functions marked as `hidden` do not
appear in the top bar menu, similarly to their behaviour in the composer
formula autocomplete assistant.

closes odoo/o-spreadsheet#3875

Task: 3810284
X-original-commit: f883740ea1d68a8ad7e8f803da691a1c92eb69c4
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>