[FIX] EditionPlugin: data validation duplicate values
Previously, when creating data validation with either a list of values
or a range of values, it resulted in an error being thrown by the
auto-complete dropdown. This occurred because the values were being used
as keys, and by definition, having duplicate keys in a list is nonsensical.

This commit addresses the problem by eliminating duplicate values from
the list.

closes odoo/o-spreadsheet#3932

Taskid: 3768392
X-original-commit: 90bd1b55f9987b51ce42d9be2ccf2bd71a541e8a
Signed-off-by: Adrien Minne (adrm) <adrm@odoo.com>
Signed-off-by: Dhrutik Patel (dhrp) <dhrp@odoo.com>