[FIX] EditionPlugin: data validation duplicate values
Previously, when creating data validation with either a list of values
or a range of values, it resulted in an error being thrown by the
auto-complete dropdown. This occurred because the values were being used
as keys, and by definition, having duplicate keys in a list is nonsensical.

This commit addresses the problem by eliminating duplicate values from
the list.

closes odoo/o-spreadsheet#3926

Taskid: 3768392
X-original-commit: c2eed35cb258ae0de1c7cc386ece59053d973c0b
Signed-off-by: Adrien Minne (adrm) <adrm@odoo.com>
Signed-off-by: Dhrutik Patel (dhrp) <dhrp@odoo.com>