[IMP] BottomBarStatistic: cache the statistics results
Currently, the statistics results for the bottom bar are recomputed at
each call, which occurs everytime we force a full rendering (i.e. when
we dispatch a command). This computaion can be very costy depending on
the size of the current selection.
This commit moves the concerned code into a dedicated store (no need to
be part of the selection plugin) and caches the results, only
invalidating them whenever the selection is updated or if a command
susceptible to affect the statistics.

Task: 3820568
Part-of: odoo/o-spreadsheet#3876