[IMP] BottomBarStatistic: Keep stats menu open and up-to-date
Currently, the statistics choice menu is closed everytime the stats
value are changed. This has two drawbacks:
- if the stats were changed following a collaborative update (e.g. change a
  cell value), the menu would close against the user will and heavy
edition of a sheet could mkae it hard to keep that menu open for long
- the current implementation was checking differences in the values
  which could lead to inconsistent behaviour (if you change selection
between cells with the same evalued content, the menu stays open, but not
if the value change)

This revision changes the "automatic closing" to be only triggered when
there are not statistics and  makes the menu item values follow the stats
changes.

closes odoo/o-spreadsheet#3876

Task: 3820568
Signed-off-by: Vincent Schippefilt (vsc) <vsc@odoo.com>