[FIX] *: Compute max/min on huge arrays
This is an extension of PR #1217. Navigators (or rather javascript
engines) have a limit to the number of function arguments [1].

We sometimes have to use Math.max and Math.min functions (which are
subject to this limitation) on every cell/column/row which means that
we are functionally limited.

E.g. If we take the limitation of Chrome, we could only process 117k
rows, which is not such a huge number if we consider the situation of a
user importing their data.

This revision introduces wrappers `largeMax` and `largeMin` which take
an  1-D array (a single argument) and find the maximum (resp. minimum)
present in that list. For arrays with les than a 100K cells, the wrapper
will still call Math.max as its implementation is most of the time
faster.

[1]: Experimentally, that limit is 117k on Google Chrome and 501k on
Firefox.

closes odoo/o-spreadsheet#3953

Task: 3802691
X-original-commit: c2f6da21963ffa5aa23ec1385f4d9b370ec88bab
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>