[IMP] BottomBarStatistics: Lazify statistic evaluation
Each time we fetch the statistics, we compute all 6 of them even though
we only need one of them to display (We need all of them when the
statistics selection menu is open).

This revision lazifies the computation of those statistics.

Benchmark
---------

In Firefox with the largeDataSet of 260k cells, averaged on 10 runs:

         _computeStatisticFnResult
-----------------------------------
before:           138  ms
after:            95.8 ms

Task: 3827324