[REF] side panel: update PanelColorPicker style
This commit changes the color picker icon to in the "design" chart
side panel by a simple circle with the color. That way we better see
the color selected, and we can differentiate between
"no color selected" and "white" (and it's prettier).

The commit also make the component more generic, it now doesn't have
a `Section` inside its template, and the text can be customized.

Task: 3836993
Part-of: odoo/o-spreadsheet#3936