[FIX]: Filters: Do not export 1-row filters to xlsx files
When exporting the data to Excel format,  the filters are processed as
if they were tables but unfortunately, Excel does not tables that span
over a single row.

As the very concept of a single row filter is not useful to the
end-user, we just skip it when exporting the data to the xlsx format.

closes odoo/o-spreadsheet#3988

Task: 3839556
X-original-commit: aa18fd20ebdf5b8e0835c5818328c6abb9d2bb0f
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>