[FIX]xlsx: replace json.stringify with deepEquals
JSON.stringify is way slower than deepEquals to compare objects and
assuming that the keys of the object are not ordered the same way, will
yield false negatives.

e.g.
```javascript
a = {'a':2,'b':4}
b = {'b': 4, 'a':2}
JSON.stringify(a) === JSON.stringify(b) // false
deepEquals(a,b) // true
```

Benchmark of `getXLSX`
======================
(averaged on 10 runs)

- On a 30k random text cells (to force heavy use of pushElement):
  Before: 153 seconds
  After: 10 seconds

- On RNG heavy sheet (patch tested in master):
  Before: 351 seconds
  After : 15 seconds

Task: 3733743
Part-of: odoo/o-spreadsheet#4015