[IMP] demo: fix random chart dataset
In the demo sheet we have a dataset with random values that is
very useful to create a chart on it. But since there is some data
adjacent to it, the other data is also selected when creating a chart
on the dataset, leading to a broken chart.

This commit moves the data a bit to isolate the chart dataset.

closes odoo/o-spreadsheet#4043

Signed-off-by: Vincent Schippefilt (vsc) <vsc@odoo.com>