[ADD] pivot: introduce spreadsheet pivot table
This commit introduces the spreadsheet pivot table. It is a new type of
pivot table that is based on values from the spreadsheet itself.

The pivot table is created by selecting a range of cells in the spreadsheet.
The first row of the range is used as the available fields, and the rest
of the range is used as the data.

The type of the data is automatically detected based on the content of
the cells in the column. It could be `date` if all the cells in the column
are dates, `boolean` if all the cells are boolean, `number` if all the
cells are numbers, or `char` otherwise.

Each field can be used as a dimension (row or column) or as a measure
(value), based on its type. The user can drag and drop the fields to
the corresponding area in the pivot table side panel, select the order,
select the granularity if the field is a date. Each measure can be
aggregated using different functions (sum, average, count, etc.).

For now, the pivot table is updated in real-time when the user changes
the range of cells in the spreadsheet. This could be heavy for large
spreadsheets, but a flag will be implemented to delay the update in a
future task (3897841).

This commit is a first version of the pivot table. Features are missing
compared to the pivot table implemented in Odoo, for example the
possibility to use `PIVOT.HEADER` and `PIVOT.VALUE` functions (and all
the features that come with it, autocompletion of formulas, positional
arguments, etc.). This will be implemented in a future task (3897857).

Task: 3748717
Part-of: odoo/o-spreadsheet#4025
Co-authored-by: rrahir <rar@odoo.com>
Co-authored-by: Pierre Rousseau <pro@odoo.com>