[FIX] charts: time axis broken with long formatted date
In the charts, we truncate the labels if they are too long
(>20 characters)[1]. But in line chart with time axis, chartJS parses
the labels to get a date. This parsing obviously fails if the label
is truncated.

This commit fixes the issue by not truncating the labels if the
axis is a time axis.

[1] e.g. a format like "12/04/2024 12:34:56 AM".

closes odoo/o-spreadsheet#4127

Task: 3857242
X-original-commit: e798934b5b4ec8e3aa7d3e6c53d4387a819e64db
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>
Signed-off-by: Adrien Minne (adrm) <adrm@odoo.com>