[FIX] chart: avoid useless chart updates
There was an `useEffect` in the `chartJS` component to update the
chartJS object on runtime change. There was 2 problems:

1) useEffect on object compare the references, so if the chart plugin
was ever changed to return copy of runtime it would break
2) the chart plugin re-build every chartRuntime on each command that
could affect a chart. This created a chart update at each UPDATE_CELL
even if the actual runtime didn't change.

This commit replace the `useEffect` dependency with a `deepEquals`

closes odoo/o-spreadsheet#4155

Task: 3697660
X-original-commit: 6eb43533daead4d4e2950d21bd514f2b84942c0e
Signed-off-by: Pierre Rousseau (pro) <pro@odoo.com>