[IMP] model: allow to batch commands in one history step
We sometime need to create a UI plugin specifically to handle a command
so that its sub-commands are batched in a single history step.

This is kind of a problem when working with stores, because then we
need both a store to handle all the business logic, and a plugin that
handle a command created specifically to batch sub-commands (looking
at you find & replace).

We can fix that easily by creating a local command `BATCH_COMMANDS`
that take a callback as argument, and that batches every command
and sub-command executed in the callback within one history step.

Task: 3870119