[FIX] borders: wrong borders on remove rows
When removing multiple headers, the borders are not shifted
correctly, leading to incorrect borders.

This happens because we modify the borders for each deleted rows
separately. But we don't sort the rows before modifying the borders.
So we first delete the row 1, then the row 2. But the row 2 is now
the row 1, since the old row 1 has been deleted. So we end up with wrong
operations.

This is solved by simply sorting the rows in reverse order before
modifying the borders.

closes odoo/o-spreadsheet#4119

Task: 3884112
X-original-commit: 9203154bee1468aba2f87b3714bef0ef197b1ea8
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>
Signed-off-by: Adrien Minne (adrm) <adrm@odoo.com>