[IMP] clipboard: insert new rows when pasting charts
Currently, when pasting a chart at the edges of a sheet, the pasted
chart is shifted so that it doesn't go out of the sheet.

This commit changes the behaviour so that now the chart is always
pasted with its top-left corner at the top-left corner of the selection,
and insert new rows/columns if needed so that the chart fits in the
sheet.

Also take the occasion to factorize the code in common between the
chart clipboard and the image clipboard into the figure clipboard.

Task: 3869372