[IMP] table: add button to edit custom style
This commit adds a button on hover on the table style preview to edit
a custom table style. The functionality was previously hidden in
a context menu, that the user couldn't know existed if he didn't
try right-clicking on the table style preview or had previous experience
on Excel.

closes odoo/o-spreadsheet#4136

Task: 3890174
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>