[IMP] clipboard: improve cut formula behaviour
Previously, when cutting a formula `=A1` in sheet1, and pasting it in
sheet2, the pasted formula would refer to `A1` in sheet2. The more
correct behaviour is to refer to `A1` in sheet1, transforming the
pasted formula to `=Sheet1!A1`.

Also when cutting a formula `=Sheet2!A1` from sheet1 to sheet2,
we can drop the prefix `Sheet2!` and only paste `=A1` in sheet2.

closes odoo/o-spreadsheet#4152

Task: 3898438
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>