[REF] clipboard: remove usage of `MOVE_RANGES`
This commit changes the clipboard to display a new command
`MOVE_REFERENCE` instead of `MOVE_RANGES`. The new command only moves
the ranges that are references to cells (in formulas in chart) instead
of modifying all the existing ranges.

There is two main reasons:

- the handling of `MOVE_RANGE` is bugged in most plugins when it
moves a range from a sheet to another. Fixing the command could lead
to broken spreadsheets, so for now we prefer to remove it.

- cut/paste was in a strange state where some operations were done
twice. For example a filter table was moved once when handling
`MOVE_RANGES` and then moved again when the cut/paste of the table
was actually done. It's better to handle everything inside the
clipboard.

Also the OT and allowDispatch of `MOVE_RANGES` was broken (it didn't
take into account that the col/row were on the targetSheetId, not on
the sheetId). This is fixed for `MOVE_REFERENCE`.

Note: when introducing dynamic tables we removed the possibility to
copy/paste filter values. It turns out that it was still somewhat
working by chance because of `MOVE_RANGES`. This is now removed.

Task: 3899595