[IMP] dashboard: automatically reorganize a sheet
This commit aims to simplify the process of creation of a dashboard
with two new features:

1) A context menu item to send figures and ranges to another sheet
2) A context menu to reorganize a sheet in a "dashboard-y" way.

Both features are implemented in the `ReorganizeSheetPlugin`. Sending
a figure/range to a sheet works by finding the bottom-most used cell
in the sheet and placing the figure/range below it. "Used" cells are
cells that contains some content/a merge/a table/overlap a figure.

Reorganizing a sheet works by:
- put the scorecards at the top of the sheet
- put the figures below the scorecards
- put the cell "clusters" (zones of contiguous cells) below them

Dev/architecture notes:

Main problematic: to reorganize cells, we could want for example to
move D1 to A1, and D1 to A1. How to do that ?
1) store D1 and A1 on "clipboards", remove content of A1 and D2, then
paste the clipboards. This doesn't work because if D1 is a formula =A1,
it won't be adapted to the new position of A1 while stored in a clipboard
2) Add a new empty sheet, and do all the moving by cut/paste to this
new sheet. At the end remove the old sheet, and rename the new sheet.
This is a bit annoying because we can end up with formulas like
=NewSheet!A1 instead of =A1, that we need to clean up. It also breaks
CTRL+Z because the activeSheetId is now the new sheet id, when CTRL+Z
the sheet is deleted and we fallback to the very first sheet
3) Add new rows at the bottom of the sheet, cut/paste into the new rows,
 and remove the old rows at the end annoying because adding new rows
 at the bottom body the size/format of the last row

We chose the option 3) because it has the less drawbacks and breaking
issues.

Task: 3869372