[FIX] gridOverlay: self-handle grid clicks
How to reproduce:

- Add a filter on a spreadsheet,
- MouseDown on the filter icon (in the grid),
- drag the mouse -> KABOOM

This issue is present since this PR[^1]. Using the behaviour of a parent
component just for the sake of sparing some code/ behaviour redundancy
seems risky and the bug introduced with [^1] is an example. This
revision considers another approach where `FilterIconOverlay` is not
aware of its parent behaviour. We simply let events bubble and both
`GridOverlay` and `Grid` take their own responsibilities.

[^1]: https://github.com/odoo/o-spreadsheet/pull/3726

closes odoo/o-spreadsheet#4266

Task: 3868838
Signed-off-by: Pierre Rousseau (pro) <pro@odoo.com>