[FIX] FigureComponent: Stop Ctrl+A propagation
Previously, when a figure was selected, And pressing CTRL + A would
select all rows and columns in the sheet, and pressing the Delete key
would delete the selected figure. This occurred because all key events
were bubbling up to the grid component.

This PR stops the propagation for the Ctrl+A key event in the figure
component.

closes odoo/o-spreadsheet#4290

Task: 3863300
X-original-commit: 03da6b422e6bb86d62238d6292ab4f960ae12a97
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>