[FIX] FigureComponent: Stop Ctrl+A propagation
Previously, when a figure was selected, And pressing CTRL + A would
select all rows and columns in the sheet, and pressing the Delete key
would delete the selected figure. This occurred because all key events
were bubbling up to the grid component.

This PR stops the propagation for the Ctrl+A key event in the figure
component.

closes odoo/o-spreadsheet#4301

Task: 3863300
X-original-commit: d0d08ec6ed5a20c1f9b7d03e96aaa4c306987f30
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>