[FIX] clipboard: cannot paste chart with deleted ranges
Before this commit, when pasting a chart with ranges in its data/labels
that were deleted between the copy and the paste, the clipboard would
traceback.

That's because `getRangeString` crashed when the sheetId does not
exist anymore. Fixed it by making the function return a string with an
invalidSheetName error rather than crashing.

Also replace the uses of `getRangeString` with `getValidRangeString`,
which will return undefined for invalid ranges so the typing forces us
to handle it properly.

closes odoo/o-spreadsheet#4052

Task: 3618758
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>