[FIX] border: bottom sheet borders removed on DELETE_ROWS
Before this commit, when we add borders at the bottom of a sheet (at very
the last rows), if we delete some rows in a position before the borders,
some/all borders disappear (depending on how many rows we deleted). The
reason being that the sheet plugin updates the total rows before we loop
through them to shift the border.

This commit fixes this behaviour by redefining the way we get the border
to shift.

closes odoo/o-spreadsheet#4604

Task: 3911695
X-original-commit: 1d1bbbdd96cd762b0f18a1cc638e637ac9b0b9e5
Signed-off-by: Mehdi Rachico (mera) <mera@odoo.com>
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>