[FIX] figures: compute correctly the snap lines in scrolled viewport
Before this fix, when working in a scrolled viewport (down or right), when resizing
a figure, the snaplines were computed without taking into account the scroll positions.

After this fix, the snaplines are computed including the scroll position.

TODO:
- add more tests that shows the snaplines are displayed correctly
- rewrite the whole "snapline" functions, see if they can be simplified
by not mixing screen positions (with offsets) to absolute positions of figures

Task: 3752042