[REF] uiPlugins: remove state observer from UI plugins
The state observer exists in core plugins to update the history with changes
that are happening during the execution of core commands, in order to undo and redo them.

After this ref, we do not give access to history to UI Plugins as they should not manage their state
though history and revisions.