[IMP] pivot: remove label of first row group by
This commit removes the label of the first row group by in the pivot table.
It was done to ease the creation of a chart from a pivot table, but it
was confusing for the user.

For the record, it was introduced here: https://github.com/odoo/odoo/pull/107220

closes odoo/o-spreadsheet#4363

Task: 3975326
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>