[IMP] charts: add chart type picker popover
Now that we start to have quite a few chart types, having a simple
dropdown to select the chart type is not very user-friendly.

This commit adds a popover to the chart type picker, which shows a
preview of each chart type. The popover is displayed instead of the
dropdown when the user clicks on the chart type `select` element.

closes odoo/o-spreadsheet#4377

Task: 3978202
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>