[FIX] ComposerStore: missing previous range
The updateComposerRange function previously assumed that a previous range would
always be found because it was only called when changing a highlight, which was
a valid assumption under past conditions.

However, this assumption is no longer accurate. There are now two scenarios
where this function can be called without a previous range:

1. With PR#3809, entering edition mode in a cell and opening the link editor
causes a traceback due to the missing previous range.
2. Opening a cell in editing mode and creating a data filter while the cell is
in edition mode also results in the same traceback.

To resolve this issue, we can return early if no update or replacement is
needed within the composer range.

closes odoo/o-spreadsheet#4480

Task: 3972392
X-original-commit: 47dd5c61911a61d4fb84af0c1bf2511e329206e3
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>
Signed-off-by: Dhrutik Patel (dhrp) <dhrp@odoo.com>