[IMP] pivot: add cog wheel menu for some actions
This revision adds a cog wheel menu to open a menu which contains the
actions "Duplicate" and "Delete" of the pivots.

To achieve this, this commit introduces two new components:
- CogWheelMenu
A components which contains a cog wheel button, which opens a menu
with actions given as props.
- PivotTitleSection
A component which contains the EditableName of the pivot and the cog
wheel menu with the actions "Duplicate" and "Delete".

Later, the EditableName will be replaced by an editable input (3964069)

closes odoo/o-spreadsheet#4375

Task: 3975491
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>