[FIX] popover: popover position with scrollbar
When opening a popover containing a scrollbar, for a single frame the
popover is wrongly positioned.

This is because we compute at the same time the popover position and
its maximum size. But setting the max size of an element can change its
dimensions (if a scrollbar appear), and changing the dimension of a
popover has an impact on its position.

closes odoo/o-spreadsheet#4434

Task: 3981551
X-original-commit: 310f874d0c9122b6006a43c92dd88a381ae87e5c
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>
Signed-off-by: Adrien Minne (adrm) <adrm@odoo.com>