[IMP] pivot_time_adapter: store number and avoid useless conversion
With this commit, the normalized value stored for a date with a
granularity of "day" is now a number instead of a string. This number
is the number of the date as stored in the spreadsheet.

This commit also simplify other adapters that handled numbers, as they
made some useless conversion.

In the pivot of the demo page (makePivotDataset) with 5k lines grouped
by date:day, the time for the evaluations drops from ± 75s to ± 15s.

Task: 3989580