[FIX] model: mark default translations as loaded
Currently the default translations function is marked as not loaded.
That means that if no custom translation method is set, the translations
will never be marked as loaded, and charts will crash because we try
to deepCopy LazyTranslatedString.

However we can't mark the default translations as loaded by default,
because otherwise top-level translations would always be translated
before we can set the translation method.

The solution is to mark the default translations as loaded only
when starting the model.

closes odoo/o-spreadsheet#4510

Task: 3999116
X-original-commit: 96506e48fd7388fe6822210876fde191edd7029c
Signed-off-by: Pierre Rousseau (pro) <pro@odoo.com>