[FIX] ChartTitle: traceback when chart title is undefined
With my recent PR#4476, creating a chart from scratch via the
top menu item results in the chart title being undefined.
This causes a traceback in the side panel since the ChartTitle
component expects the title prop to be a string.

This commit fixes the issue by providing a default props title
value of an empty string.

closes odoo/o-spreadsheet#4520

Task: 0
Signed-off-by: Adrien Minne (adrm) <adrm@odoo.com>