[REV] chart: avoid useless chart updates
This reverts commit 6eb43533d.

It turns out that checking the deep equality of runtime while ignoring
functions is not a good fix to avoid useless chart updates. The problem
is that some runtime changes apply only to the callbacks (eg. changing
the dataset format only change the ticks callback).

The only real alternative would be to add the variables that are used
in the callbacks to the runtime, so the deepEquals would work. But this
is very error prone: we'll 100% forget to add a variable at some point.

So we will accept the useless updates for now, until we see a real
performance issue.

closes odoo/o-spreadsheet#4602

Task: 4029016
X-original-commit: ef4c70d651189f8d6009c60396fd2e848cec1d4b
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>
Signed-off-by: Adrien Minne (adrm) <adrm@odoo.com>