[PERF] evaluation: invalidate spreading in batch
When spreading the result of an array formula, we invalidate formulas
depending on the spread zone. Currently, we look for the dependencies
one position in the spread zone at a time.

It's much more efficient to group all those individual positions in zones
and call `this.getCellsDependingOn(...)` with those zones.

This commit reduces the time to evaluate the large array formula
dataset from 60+ seconds to ~36 seconds.

Task: 4036899
X-original-commit: 565dbdc9d8835b894acfd32059f6f1fa7edc6172
Part-of: odoo/o-spreadsheet#4622
Signed-off-by: Vincent Schippefilt (vsc) <vsc@odoo.com>
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>