[IMP] composer: stop composition on blur
ATM the composer do not lose the DOM focus if the composer is blurred.
At the next render, the composer will take back the DOM focus and
continue the edition.

This is fine for the grid/topbar composer, but it really is bad for a
standalone composer in the side panel. After editing the composer, click
on the button save of the panel ? You didn't actually confirm the
composer so you don't have the changes. Click on another input ?
You didn't confirm either. And at the next render, the focus will
randomly leave your input to go back to the composer.

Functionally we really need to stop the composition on blur event.
But some blur events are expected by the composer: if the user click
on the grid on on another sheet to select a range, the composer will
receive a blur event but it should not stop the composition.

The (wonky) idea to solve the issue is to check the relatedTarget of
the blur event (the new focused element). We will put an attribute
`composerFocusableElement` on the grid & sheet elements that should
not stop the composition. If the relatedTarget is one of these elements,
we will not stop the composition.

Task: 4045550
Part-of: odoo/o-spreadsheet#4638
Signed-off-by: Pierre Rousseau (pro) <pro@odoo.com>