[REF] compiler: compile no longer throws any errors
Currently `compile` and `compileTokens` throw errors if there's
a syntax error in the formula.

It's a frequent source of errors and traceback because it's easy to forget
to add a try/catch around the call.

Even if the developper doesn't forget, the try/catch pattern is repeated
at every call site.

With this commit, compile can no longer throw any errors. It always returns
a valid compiled formula. In case of a syntax error, the `execute`
returns the compilation error at the evaluation time. Errors in
`execute` are expected and handled in the evaluation.

A boolean is added to the compiled formula object indicating if the formula
was correctly compiled or not. It's usefull if we want to warn the user
if he created a wrong formula and reject the input.

closes odoo/o-spreadsheet#4652

Task: 4049824
Signed-off-by: Pierre Rousseau (pro) <pro@odoo.com>