[FIX] pivot: correctly define default measure aggregator
Measure aggregator are undefined by default when inserting a measure
in a pivot. This leads to a strange state with a "dynamic" aggregator,
if the field change from a integer to string field, the aggregator
will automatically switch from `sum` to `count`.

This commit fixes this issue by defining a default aggregator when
inserting a measure from the side panel.

closes odoo/o-spreadsheet#4658

Task: 4052502
X-original-commit: c3a9445e21245878b24dc087e572e452424c0e09
Signed-off-by: Pierre Rousseau (pro) <pro@odoo.com>
Signed-off-by: Adrien Minne (adrm) <adrm@odoo.com>