[FIX] pivot: correctly define default measure aggregator
Measure aggregator are undefined by default when inserting a measure
in a pivot. This leads to a strange state with a "dynamic" aggregator,
if the field change from a integer to string field, the aggregator
will automatically switch from `sum` to `count`.

This commit fixes this issue by defining a default aggregator when
inserting a measure from the side panel.

closes odoo/o-spreadsheet#4654

Task: 4052502
Signed-off-by: Pierre Rousseau (pro) <pro@odoo.com>