[IMP] pivot: support multiple measures
This commit adds support for adding the same field as measure multiple
times in the pivot. It's useful when you want to show the same measure
with different aggregator functions in the same pivot. Or, in the
future, the same measure with the same aggregate but shown in different
way.

closes odoo/o-spreadsheet#4684

Task: 4001750
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>