[PERF] evaluation: store spreading relation by zones
Evaluating the large vectorized dataset is now as fast as evaluating the
large formula dataset.
Which means loading the entire vectorized model is ~4x faster since it
doesn't need to import 260k cells

This commit changes the implementation of the spreading relations
data structures.

The main objective is to work with spilled zones instead
of each individual positions in spilled zones.

Before, if a formula spilled on a zone of 10k cells, we added each of these
10k positions in the spreading relations data structure. Now we only add a
single zone (spanning those 10k cells).

For an extreme example: on the vectorized large dataset,
there a 26 spilled zones (26 array formulas) but there are
260k cell positions.

Insertions are faster:
- before: O(Z*C)
- after: O(Z*log Z)

where Z is the number of array formulas and C is the number
of cells in spilled zones. Note that C can be very large !

This comes with a trade-off though: looking up the array formula
spilling on a given position is slower:
- before: O(1)
- after: O(log Z)

It's still faster overall in most situations, because we can now lookup for
an entire zone at the same cost rather than position by position.
It's only slower when we actually need to lookup for many individual
positions.

Some measures:

*Evaluate large formula dataset:
no impact

*Evaluate RNG's "BE Timesheet" dashboard:
no impact

*Evaluate large vectorized dataset:
- before: 550ms
- after:  210ms (-61%)

*Copy all cells in the large vectorized dataset:
- before: 470ms
- after:  631ms (+34%)

* Copy all cells in the large formula dataste:
no impact

Task: 4116073
Part-of: odoo/o-spreadsheet#4692
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>