[IMP] cf panel: improve layout on panel resize
The conditional formatting panel layout wasn't very good when
the panel was resized. The interesting inputs (the values) stayed
small while the labels and other select were taking all the space.

This was a bit of a problem when trying to edit a composer in this
panel, it was hard to see the values.

Task: 4045550
Part-of: odoo/o-spreadsheet#4638
Signed-off-by: Pierre Rousseau (pro) <pro@odoo.com>