[IMP] pivot: allow to partially reload pivot runtime
Currently, when a pivot is updated via an UPDATE_PIVOT command, the
entire pivot object is recreated from scratch. As a result, all the
setup of the pivot is executed (e.g. RPCs in Odoo, fields extraction
from dataSet, etc.) with each update. This is necessary when the pivot
'structure' changes (columns, rows, measures, etc.), but it is
unnecessary when only the name is updated. This will become an issue
with renaming pivot measures, adding calculated fields, and measures
shown as 'x', as these three operations do not require the setup to be
re-executed.

It's already useful in this PR to avoid re-extracting the fields from
the dataSet if it hasn't changed.

closes odoo/o-spreadsheet#4674

Task: 4070141
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>